Crate google-youtube3

This documentation was generated from YouTube crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

activities (insert and list)

channel banners (insert)

channel sections (delete, insert, list and update)

channels (list and update)

guide categories (list)

i18n languages (list)

i18n regions (list)

live broadcasts (bind, control, delete, insert, list, transition and update)

live streams (delete, insert, list and update)

playlist items (delete, insert, list and update)

playlists (delete, insert, list and update)

search (list)

subscriptions (delete, insert and list)

thumbnails (set)

video categories (list)

videos (delete, get rating, insert, list, rate and update)

watermarks (set and unset)

Upload supported by ...

set watermarks

insert channel banners

set thumbnails

insert videos

Subscription supported by ...

list playlist items

Everything else about the YouTube v3 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.live_broadcasts().control(...).doit()
let r = hub.live_broadcasts().insert(...).doit()
let r = hub.live_broadcasts().list(...).doit()
let r = hub.live_broadcasts().transition(...).doit()
let r = hub.live_broadcasts().update(...).doit()
let r = hub.live_broadcasts().delete(...).doit()
let r = hub.live_broadcasts().bind(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-youtube3 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-youtube3" as youtube3;
use youtube3::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.live_broadcasts().list("part")
             .page_token("justo")
             .on_behalf_of_content_owner_channel("amet.")
             .on_behalf_of_content_owner("erat")
             .mine(true)
             .max_results(92)
             .id("nonumy")
             .broadcast_status("dolores")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

AccessPolicy
Rights management policy for YouTube resources.

This type is not used in any activity, and only used as part of another schema.

Activity
An activity resource contains information about an action that a particular channel, or user, has taken on YouTube.The actions reported in activity feeds include rating a video, sharing a video, marking a video as a favorite, commenting on a video, uploading a video, and so forth. Each activity resource identifies the type of action, the channel associated with the action, and the resource(s) associated with the action, such as the video that was rated or uploaded.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert activities (request|response)

ActivityContentDetails
Details about the content of an activity: the video that was shared, the channel that was subscribed to, etc.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsBulletin
Details about a channel bulletin post.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsChannelItem
Details about a resource which was added to a channel.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsComment
Information about a resource that received a comment.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsFavorite
Information about a video that was marked as a favorite video.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsLike
Information about a resource that received a positive (like) rating.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsPlaylistItem
Information about a new playlist item.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsPromotedItem
Details about a resource which is being promoted.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsRecommendation
Information that identifies the recommended resource.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsSocial
Details about a social network post.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsSubscription
Information about a channel that a user subscribed to.

This type is not used in any activity, and only used as part of another schema.

ActivityContentDetailsUpload
Information about the uploaded video.

This type is not used in any activity, and only used as part of another schema.

ActivityInsertMethodBuilder
Posts a bulletin for a specific channel. (The user submitting the request must be authorized to act on the channel's behalf.)

Note: Even though an activity resource can contain information about actions like a user rating a video or marking a video as a favorite, you need to use other API methods to generate those activity resources. For example, you would use the API's videos.rate() method to rate a video and the playlistItems.insert() method to mark a video as a favorite.

ActivityListMethodBuilder
Returns a list of channel activity events that match the request criteria. For example, you can retrieve events associated with a particular channel, events associated with the user's subscriptions and Google+ friends, or the YouTube home page feed, which is customized for each user.

ActivityListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list activities (response)

ActivityMethodsBuilder
A builder providing access to all methods supported on activity resources. It is not used directly, but through the YouTube hub.

ActivitySnippet
Basic details about an activity, including title, description, thumbnails, activity type and group.

This type is not used in any activity, and only used as part of another schema.

CdnSettings
Brief description of the live stream cdn settings.

This type is not used in any activity, and only used as part of another schema.

Channel
A channel resource contains information about a YouTube channel.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list channels (none)

update channels (request|response)

ChannelAuditDetails
The auditDetails object encapsulates channel data that is relevant for YouTube Partners during the audit process.

This type is not used in any activity, and only used as part of another schema.

ChannelBannerInsertMethodBuilder
Uploads a channel banner image to YouTube. This method represents the first two steps in a three-step process to update the banner image for a channel:

Call the channelBanners.insert method to upload the binary image data to YouTube. The image must have a 16:9 aspect ratio and be at least 2120x1192 pixels.

Extract the url property's value from the response that the API returns for step 1.

Call the channels.update method to update the channel's branding settings. Set the brandingSettings.image.bannerExternalUrl property's value to the URL obtained in step 2.

ChannelBannerMethodsBuilder
A builder providing access to all methods supported on channelBanner resources. It is not used directly, but through the YouTube hub.

ChannelBannerResource
A channel banner returned as the response to a channel_banner.insert call.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert channel banners (request|response)

ChannelBrandingSettings
Branding properties of a YouTube channel.

This type is not used in any activity, and only used as part of another schema.

ChannelContentDetails
Details about the content of a channel.

This type is not used in any activity, and only used as part of another schema.

ChannelContentDetailsRelatedPlaylists
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ChannelContentOwnerDetails
The contentOwnerDetails object encapsulates channel data that is relevant for YouTube Partners linked with the channel.

This type is not used in any activity, and only used as part of another schema.

ChannelConversionPing
Pings that the app shall fire (authenticated by biscotti cookie). Each ping has a context, in which the app must fire the ping, and a url identifying the ping.

This type is not used in any activity, and only used as part of another schema.

ChannelConversionPings
The conversionPings object encapsulates information about conversion pings that need to be respected by the channel.

This type is not used in any activity, and only used as part of another schema.

ChannelListMethodBuilder
Returns a collection of zero or more channel resources that match the request criteria.

ChannelListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list channels (response)

ChannelLocalization
Channel localization setting

This type is not used in any activity, and only used as part of another schema.

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the YouTube hub.

ChannelSection
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete channel sections (none)

update channel sections (request|response)

list channel sections (none)

insert channel sections (request|response)

ChannelSectionContentDetails
Details about a channelsection, including playlists and channels.

This type is not used in any activity, and only used as part of another schema.

ChannelSectionDeleteMethodBuilder
Deletes a channelSection.

ChannelSectionInsertMethodBuilder
Adds a channelSection for the authenticated user's channel.

ChannelSectionListMethodBuilder
Returns channelSection resources that match the API request criteria.

ChannelSectionListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list channel sections (response)

ChannelSectionLocalization
ChannelSection localization setting

This type is not used in any activity, and only used as part of another schema.

ChannelSectionMethodsBuilder
A builder providing access to all methods supported on channelSection resources. It is not used directly, but through the YouTube hub.

ChannelSectionSnippet
Basic details about a channel section, including title, style and position.

This type is not used in any activity, and only used as part of another schema.

ChannelSectionUpdateMethodBuilder
Update a channelSection.

ChannelSettings
Branding properties for the channel view.

This type is not used in any activity, and only used as part of another schema.

ChannelSnippet
Basic details about a channel, including title, description and thumbnails.

This type is not used in any activity, and only used as part of another schema.

ChannelStatistics
Statistics about a channel: number of subscribers, number of videos in the channel, etc.

This type is not used in any activity, and only used as part of another schema.

ChannelStatus
JSON template for the status part of a channel.

This type is not used in any activity, and only used as part of another schema.

ChannelTopicDetails
Freebase topic information related to the channel.

This type is not used in any activity, and only used as part of another schema.

ChannelUpdateMethodBuilder
Updates a channel's metadata.

ContentRating
Ratings schemes. The country-specific ratings are mostly for movies and shows. NEXT_ID: 65

This type is not used in any activity, and only used as part of another schema.

GeoPoint
Geographical coordinates of a point, in WGS84.

This type is not used in any activity, and only used as part of another schema.

GuideCategory
A guideCategory resource identifies a category that YouTube algorithmically assigns based on a channel's content or other indicators, such as the channel's popularity. The list is similar to video categories, with the difference being that a video's uploader can assign a video category but only YouTube can assign a channel category.

This type is not used in any activity, and only used as part of another schema.

GuideCategoryListMethodBuilder
Returns a list of categories that can be associated with YouTube channels.

GuideCategoryListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list guide categories (response)

GuideCategoryMethodsBuilder
A builder providing access to all methods supported on guideCategory resources. It is not used directly, but through the YouTube hub.

GuideCategorySnippet
Basic details about a guide category.

This type is not used in any activity, and only used as part of another schema.

I18nLanguage
An i18nLanguage resource identifies a UI language currently supported by YouTube.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list i18n languages (none)

I18nLanguageListMethodBuilder
Returns a list of supported languages.

I18nLanguageListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list i18n languages (response)

I18nLanguageMethodsBuilder
A builder providing access to all methods supported on i18nLanguage resources. It is not used directly, but through the YouTube hub.

I18nLanguageSnippet
Basic details about an i18n language, such as language code and human-readable name.

This type is not used in any activity, and only used as part of another schema.

I18nRegion
A i18nRegion resource identifies a region where YouTube is available.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list i18n regions (none)

I18nRegionListMethodBuilder
Returns a list of supported regions.

I18nRegionListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list i18n regions (response)

I18nRegionMethodsBuilder
A builder providing access to all methods supported on i18nRegion resources. It is not used directly, but through the YouTube hub.

I18nRegionSnippet
Basic details about an i18n region, such as region code and human-readable name.

This type is not used in any activity, and only used as part of another schema.

ImageSettings
Branding properties for images associated with the channel.

This type is not used in any activity, and only used as part of another schema.

IngestionInfo
Describes information necessary for ingesting an RTMP or an HTTP stream.

This type is not used in any activity, and only used as part of another schema.

InvideoBranding
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set watermarks (request)

InvideoPosition
Describes the spatial position of a visual widget inside a video. It is a union of various position types, out of which only will be set one.

This type is not used in any activity, and only used as part of another schema.

InvideoPromotion
Describes an invideo promotion campaign consisting of multiple promoted items. A campaign belongs to a single channel_id.

This type is not used in any activity, and only used as part of another schema.

InvideoTiming
Describes a temporal position of a visual widget inside a video.

This type is not used in any activity, and only used as part of another schema.

LanguageTag
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LiveBroadcast
A liveBroadcast resource represents an event that will be streamed, via live video, on YouTube.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

control live broadcasts (response)

insert live broadcasts (request|response)

list live broadcasts (none)

transition live broadcasts (response)

update live broadcasts (request|response)

delete live broadcasts (none)

bind live broadcasts (response)

LiveBroadcastBindMethodBuilder
Binds a YouTube broadcast to a stream or removes an existing binding between a broadcast and a stream. A broadcast can only be bound to one video stream.

LiveBroadcastContentDetails
Detailed settings of a broadcast.

This type is not used in any activity, and only used as part of another schema.

LiveBroadcastControlMethodBuilder
Controls the settings for a slate that can be displayed in the broadcast stream.

LiveBroadcastDeleteMethodBuilder
Deletes a broadcast.

LiveBroadcastInsertMethodBuilder
Creates a broadcast.

LiveBroadcastListMethodBuilder
Returns a list of YouTube broadcasts that match the API request parameters.

LiveBroadcastListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list live broadcasts (response)

LiveBroadcastMethodsBuilder
A builder providing access to all methods supported on liveBroadcast resources. It is not used directly, but through the YouTube hub.

LiveBroadcastSnippet
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LiveBroadcastStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LiveBroadcastTransitionMethodBuilder
Changes the status of a YouTube live broadcast and initiates any processes associated with the new status. For example, when you transition a broadcast's status to testing, YouTube starts to transmit video to that broadcast's monitor stream. Before calling this method, you should confirm that the value of the status.streamStatus property for the stream bound to your broadcast is active.

LiveBroadcastUpdateMethodBuilder
Updates a broadcast. For example, you could modify the broadcast settings defined in the liveBroadcast resource's contentDetails object.

LiveStream
A live stream describes a live ingestion point.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete live streams (none)

update live streams (request|response)

list live streams (none)

insert live streams (request|response)

LiveStreamContentDetails
Detailed settings of a stream.

This type is not used in any activity, and only used as part of another schema.

LiveStreamDeleteMethodBuilder
Deletes a video stream.

LiveStreamInsertMethodBuilder
Creates a video stream. The stream enables you to send your video to YouTube, which can then broadcast the video to your audience.

LiveStreamListMethodBuilder
Returns a list of video streams that match the API request parameters.

LiveStreamListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list live streams (response)

LiveStreamMethodsBuilder
A builder providing access to all methods supported on liveStream resources. It is not used directly, but through the YouTube hub.

LiveStreamSnippet
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LiveStreamStatus
Brief description of the live stream status.

This type is not used in any activity, and only used as part of another schema.

LiveStreamUpdateMethodBuilder
Updates a video stream. If the properties that you want to change cannot be updated, then you need to create a new stream with the proper settings.

LocalizedProperty
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LocalizedString
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

MonitorStreamInfo
Settings and Info of the monitor stream

This type is not used in any activity, and only used as part of another schema.

PageInfo
Paging details for lists of resources, including total number of items available and number of resources returned in a single page.

This type is not used in any activity, and only used as part of another schema.

Playlist
A playlist resource represents a YouTube playlist. A playlist is a collection of videos that can be viewed sequentially and shared with other users. A playlist can contain up to 200 videos, and YouTube does not limit the number of playlists that each user creates. By default, playlists are publicly visible to other users, but playlists can be public or private.

YouTube also uses playlists to identify special collections of videos for a channel, such as:
- uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information. You can retrieve the playlist IDs for each of these lists from the channel resource for a given channel.

You can then use the playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the playlistItems.insert and playlistItems.delete methods.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert playlists (request|response)

delete playlists (none)

list playlists (none)

update playlists (request|response)

PlaylistContentDetails
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PlaylistDeleteMethodBuilder
Deletes a playlist.

PlaylistInsertMethodBuilder
Creates a playlist.

PlaylistItem
A playlistItem resource identifies another resource, such as a video, that is included in a playlist. In addition, the playlistItem resource contains details about the included resource that pertain specifically to how that resource is used in that playlist.

YouTube uses playlists to identify special collections of videos for a channel, such as:
- uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information.

You can retrieve the playlist IDs for each of these lists from the channel resource for a given channel. You can then use the playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the playlistItems.insert and playlistItems.delete methods. For example, if a user gives a positive rating to a video, you would insert that video into the liked videos playlist for that user's channel.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert playlist items (request|response)

update playlist items (request|response)

list playlist items (none)

delete playlist items (none)

PlaylistItemContentDetails
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PlaylistItemDeleteMethodBuilder
Deletes a playlist item.

PlaylistItemInsertMethodBuilder
Adds a resource to a playlist.

PlaylistItemListMethodBuilder
Returns a collection of playlist items that match the API request parameters. You can retrieve all of the playlist items in a specified playlist or retrieve one or more playlist items by their unique IDs.

PlaylistItemListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list playlist items (response)

PlaylistItemMethodsBuilder
A builder providing access to all methods supported on playlistItem resources. It is not used directly, but through the YouTube hub.

PlaylistItemSnippet
Basic details about a playlist, including title, description and thumbnails.

This type is not used in any activity, and only used as part of another schema.

PlaylistItemStatus
Information about the playlist item's privacy status.

This type is not used in any activity, and only used as part of another schema.

PlaylistItemUpdateMethodBuilder
Modifies a playlist item. For example, you could update the item's position in the playlist.

PlaylistListMethodBuilder
Returns a collection of playlists that match the API request parameters. For example, you can retrieve all playlists that the authenticated user owns, or you can retrieve one or more playlists by their unique IDs.

PlaylistListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list playlists (response)

PlaylistLocalization
Playlist localization setting

This type is not used in any activity, and only used as part of another schema.

PlaylistMethodsBuilder
A builder providing access to all methods supported on playlist resources. It is not used directly, but through the YouTube hub.

PlaylistPlayer
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PlaylistSnippet
Basic details about a playlist, including title, description and thumbnails.

This type is not used in any activity, and only used as part of another schema.

PlaylistStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PlaylistUpdateMethodBuilder
Modifies a playlist. For example, you could change a playlist's title, description, or privacy status.

PromotedItem
Describes a single promoted item.

This type is not used in any activity, and only used as part of another schema.

PromotedItemId
Describes a single promoted item id. It is a union of various possible types.

This type is not used in any activity, and only used as part of another schema.

PropertyValue
A pair Property / Value.

This type is not used in any activity, and only used as part of another schema.

ResourceId
A resource id is a generic reference that points to another YouTube resource.

This type is not used in any activity, and only used as part of another schema.

SearchListMethodBuilder
Returns a collection of search results that match the query parameters specified in the API request. By default, a search result set identifies matching video, channel, and playlist resources, but you can also configure queries to only retrieve a specific type of resource.

SearchListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list search (response)

SearchMethodsBuilder
A builder providing access to all methods supported on search resources. It is not used directly, but through the YouTube hub.

SearchResult
A search result contains information about a YouTube video, channel, or playlist that matches the search parameters specified in an API request. While a search result points to a uniquely identifiable resource, like a video, it does not have its own persistent data.

This type is not used in any activity, and only used as part of another schema.

SearchResultSnippet
Basic details about a search result, including title, description and thumbnails of the item referenced by the search result.

This type is not used in any activity, and only used as part of another schema.

Subscription
A subscription resource contains information about a YouTube user subscription. A subscription notifies a user when new videos are added to a channel or when another user takes one of several actions on YouTube, such as uploading a video, rating a video, or commenting on a video.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert subscriptions (request|response)

list subscriptions (none)

delete subscriptions (none)

SubscriptionContentDetails
Details about the content to witch a subscription refers.

This type is not used in any activity, and only used as part of another schema.

SubscriptionDeleteMethodBuilder
Deletes a subscription.

SubscriptionInsertMethodBuilder
Adds a subscription for the authenticated user's channel.

SubscriptionListMethodBuilder
Returns subscription resources that match the API request criteria.

SubscriptionListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list subscriptions (response)

SubscriptionMethodsBuilder
A builder providing access to all methods supported on subscription resources. It is not used directly, but through the YouTube hub.

SubscriptionSnippet
Basic details about a subscription, including title, description and thumbnails of the subscribed item.

This type is not used in any activity, and only used as part of another schema.

SubscriptionSubscriberSnippet
Basic details about a subscription's subscriber including title, description, channel ID and thumbnails.

This type is not used in any activity, and only used as part of another schema.

Thumbnail
A thumbnail is an image representing a YouTube resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set thumbnails (none)

ThumbnailDetails
Internal representation of thumbnails for a YouTube resource.

This type is not used in any activity, and only used as part of another schema.

ThumbnailMethodsBuilder
A builder providing access to all methods supported on thumbnail resources. It is not used directly, but through the YouTube hub.

ThumbnailSetMethodBuilder
Uploads a custom video thumbnail to YouTube and sets it for a video.

ThumbnailSetResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set thumbnails (response)

TokenPagination
Stub token pagination template to suppress results.

This type is not used in any activity, and only used as part of another schema.

Video
A video resource represents a YouTube video.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rate videos (none)

get rating videos (none)

list videos (none)

insert videos (request|response)

update videos (request|response)

delete videos (none)

VideoAgeGating
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

VideoCategory
A videoCategory resource identifies a category that has been or could be associated with uploaded videos.

This type is not used in any activity, and only used as part of another schema.

VideoCategoryListMethodBuilder
Returns a list of categories that can be associated with YouTube videos.

VideoCategoryListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list video categories (response)

VideoCategoryMethodsBuilder
A builder providing access to all methods supported on videoCategory resources. It is not used directly, but through the YouTube hub.

VideoCategorySnippet
Basic details about a video category, such as its localized title.

This type is not used in any activity, and only used as part of another schema.

VideoContentDetails
Details about the content of a YouTube Video.

This type is not used in any activity, and only used as part of another schema.

VideoContentDetailsRegionRestriction
DEPRECATED Region restriction of the video.

This type is not used in any activity, and only used as part of another schema.

VideoConversionPing
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

VideoConversionPings
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

VideoDeleteMethodBuilder
Deletes a YouTube video.

VideoFileDetails
Describes original video file properties, including technical details about audio and video streams, but also metadata information like content length, digitization time, or geotagging information.

This type is not used in any activity, and only used as part of another schema.

VideoFileDetailsAudioStream
Information about an audio stream.

This type is not used in any activity, and only used as part of another schema.

VideoFileDetailsVideoStream
Information about a video stream.

This type is not used in any activity, and only used as part of another schema.

VideoGetRatingMethodBuilder
Retrieves the ratings that the authorized user gave to a list of specified videos.

VideoGetRatingResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get rating videos (response)

VideoInsertMethodBuilder
Uploads a video to YouTube and optionally sets the video's metadata.

VideoListMethodBuilder
Returns a list of videos that match the API request parameters.

VideoListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list videos (response)

VideoLiveStreamingDetails
Details about the live streaming metadata.

This type is not used in any activity, and only used as part of another schema.

VideoLocalization
Localized versions of certain video properties (e.g. title).

This type is not used in any activity, and only used as part of another schema.

VideoMethodsBuilder
A builder providing access to all methods supported on video resources. It is not used directly, but through the YouTube hub.

VideoMonetizationDetails
Details about monetization of a YouTube Video.

This type is not used in any activity, and only used as part of another schema.

VideoPlayer
Player to be used for a video playback.

This type is not used in any activity, and only used as part of another schema.

VideoProcessingDetails
Describes processing status and progress and availability of some other Video resource parts.

This type is not used in any activity, and only used as part of another schema.

VideoProcessingDetailsProcessingProgress
Video processing progress and completion time estimate.

This type is not used in any activity, and only used as part of another schema.

VideoProjectDetails
Project specific details about the content of a YouTube Video.

This type is not used in any activity, and only used as part of another schema.

VideoRateMethodBuilder
Add a like or dislike rating to a video or remove a rating from a video.

VideoRating
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

VideoRecordingDetails
Recording information associated with the video.

This type is not used in any activity, and only used as part of another schema.

VideoSnippet
Basic details about a video, including title, description, uploader, thumbnails and category.

This type is not used in any activity, and only used as part of another schema.

VideoStatistics
Statistics about the video, such as the number of times the video was viewed or liked.

This type is not used in any activity, and only used as part of another schema.

VideoStatus
Basic details about a video category, such as its localized title.

This type is not used in any activity, and only used as part of another schema.

VideoSuggestions
Specifies suggestions on how to improve video content, including encoding hints, tag suggestions, and editor suggestions.

This type is not used in any activity, and only used as part of another schema.

VideoSuggestionsTagSuggestion
A single tag suggestion with it's relevance information.

This type is not used in any activity, and only used as part of another schema.

VideoTopicDetails
Freebase topic information related to the video.

This type is not used in any activity, and only used as part of another schema.

VideoUpdateMethodBuilder
Updates a video's metadata.

WatchSettings
Branding properties for the watch.

This type is not used in any activity, and only used as part of another schema.

WatermarkMethodsBuilder
A builder providing access to all methods supported on watermark resources. It is not used directly, but through the YouTube hub.

WatermarkSetMethodBuilder
Uploads a watermark image to YouTube and sets it for a channel.

WatermarkUnsetMethodBuilder
Deletes a watermark.

YouTube
Central instance to access all YouTube related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.