Struct google_youtube3::api::LiveBroadcastContentDetails

pub struct LiveBroadcastContentDetails {
    pub bound_stream_id: Option<String>,
    pub bound_stream_last_update_time_ms: Option<String>,
    pub closed_captions_type: Option<String>,
    pub enable_auto_start: Option<bool>,
    pub enable_auto_stop: Option<bool>,
    pub enable_closed_captions: Option<bool>,
    pub enable_content_encryption: Option<bool>,
    pub enable_dvr: Option<bool>,
    pub enable_embed: Option<bool>,
    pub enable_low_latency: Option<bool>,
    pub latency_preference: Option<String>,
    pub mesh: Option<String>,
    pub monitor_stream: Option<MonitorStreamInfo>,
    pub projection: Option<String>,
    pub record_from_start: Option<bool>,
    pub start_with_slate: Option<bool>,
    pub stereo_layout: Option<String>,
}

Detailed settings of a broadcast.

This type is not used in any activity, and only used as part of another schema.

Fields
bound_stream_id: Option<String>

This value uniquely identifies the live stream bound to the broadcast.

bound_stream_last_update_time_ms: Option<String>

The date and time that the live stream referenced by boundStreamId was last updated.

closed_captions_type: Option<String>

no description provided

enable_auto_start: Option<bool>

This setting indicates whether auto start is enabled for this broadcast. The default value for this property is false. This setting can only be used by Events.

enable_auto_stop: Option<bool>

This setting indicates whether auto stop is enabled for this broadcast. The default value for this property is false. This setting can only be used by Events.

enable_closed_captions: Option<bool>

This setting indicates whether HTTP POST closed captioning is enabled for this broadcast. The ingestion URL of the closed captions is returned through the liveStreams API. This is mutually exclusive with using the closed_captions_type property, and is equivalent to setting closed_captions_type to CLOSED_CAPTIONS_HTTP_POST.

enable_content_encryption: Option<bool>

This setting indicates whether YouTube should enable content encryption for the broadcast.

enable_dvr: Option<bool>

This setting determines whether viewers can access DVR controls while watching the video. DVR controls enable the viewer to control the video playback experience by pausing, rewinding, or fast forwarding content. The default value for this property is true. Important: You must set the value to true and also set the enableArchive property's value to true if you want to make playback available immediately after the broadcast ends.

enable_embed: Option<bool>

This setting indicates whether the broadcast video can be played in an embedded player. If you choose to archive the video (using the enableArchive property), this setting will also apply to the archived video.

enable_low_latency: Option<bool>

Indicates whether this broadcast has low latency enabled.

latency_preference: Option<String>

If both this and enable_low_latency are set, they must match. LATENCY_NORMAL should match enable_low_latency=false LATENCY_LOW should match enable_low_latency=true LATENCY_ULTRA_LOW should have enable_low_latency omitted.

mesh: Option<String>

The mesh for projecting the video if projection is mesh. The mesh value must be a UTF-8 string containing the base-64 encoding of 3D mesh data that follows the Spherical Video V2 RFC specification for an mshp box, excluding the box size and type but including the following four reserved zero bytes for the version and flags.

monitor_stream: Option<MonitorStreamInfo>

The monitorStream object contains information about the monitor stream, which the broadcaster can use to review the event content before the broadcast stream is shown publicly.

projection: Option<String>

The projection format of this broadcast. This defaults to rectangular.

record_from_start: Option<bool>

Automatically start recording after the event goes live. The default value for this property is true. Important: You must also set the enableDvr property's value to true if you want the playback to be available immediately after the broadcast ends. If you set this property's value to true but do not also set the enableDvr property to true, there may be a delay of around one day before the archived video will be available for playback.

start_with_slate: Option<bool>

This setting indicates whether the broadcast should automatically begin with an in-stream slate when you update the broadcast's status to live. After updating the status, you then need to send a liveCuepoints.insert request that sets the cuepoint's eventState to end to remove the in-stream slate and make your broadcast stream visible to viewers.

stereo_layout: Option<String>

The 3D stereo layout of this broadcast. This defaults to mono.