Struct google_youtube3::api::LiveChatModeratorDeleteCall

pub struct LiveChatModeratorDeleteCall<'a, S>
where
    S: 'a,
{ /* private fields */ }

Deletes a chat moderator.

A builder for the delete method supported by a liveChatModerator resource. It is not used directly, but through a LiveChatModeratorMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.live_chat_moderators().delete("id")
             .doit().await;

Implementations

impl<'a, S> LiveChatModeratorDeleteCall<'a, S>
where S: Service<Uri> + Clone + Send + Sync + 'static, S::Response: Connection + AsyncRead + AsyncWrite + Send + Unpin + 'static, S::Future: Send + Unpin + 'static, S::Error: Into<Box<dyn StdError + Send + Sync>>,

pub async fn doit(self) -> Result<Response<Body>>

Perform the operation you have build so far.

pub fn id(self, new_value: &str) -> LiveChatModeratorDeleteCall<'a, S>

Sets the id query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn delegate( self, new_value: &'a mut dyn Delegate ) -> LiveChatModeratorDeleteCall<'a, S>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

                  It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> LiveChatModeratorDeleteCall<'a, S>
where T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

$.xgafv (query-string) - V1 error format.

access_token (query-string) - OAuth access token.

alt (query-string) - Data format for response.

callback (query-string) - JSONP

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

uploadType (query-string) - Legacy upload protocol for media (e.g. “media”, “multipart”).

upload_protocol (query-string) - Upload protocol for media (e.g. “raw”, “multipart”).

pub fn add_scope<St>(self, scope: St) -> LiveChatModeratorDeleteCall<'a, S>
where St: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead of the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.

pub fn add_scopes<I, St>(self, scopes: I) -> LiveChatModeratorDeleteCall<'a, S>
where I: IntoIterator<Item = St>, St: AsRef<str>,

Identifies the authorization scope(s) for the method you are building.

See Self::add_scope() for details.

pub fn clear_scopes(self) -> LiveChatModeratorDeleteCall<'a, S>

Removes all scopes, and no default scope will be used either. In this case, you have to specify your API-key using the key parameter (see Self::param() for details).