Struct google_youtube3::api::Video

pub struct Video {

    pub age_gating: Option<VideoAgeGating>,
    pub content_details: Option<VideoContentDetails>,
    pub etag: Option<String>,
    pub file_details: Option<VideoFileDetails>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub live_streaming_details: Option<VideoLiveStreamingDetails>,
    pub localizations: Option<HashMap<String, VideoLocalization>>,
    pub monetization_details: Option<VideoMonetizationDetails>,
    pub player: Option<VideoPlayer>,
    pub processing_details: Option<VideoProcessingDetails>,
    pub project_details: Option<VideoProjectDetails>,
    pub recording_details: Option<VideoRecordingDetails>,
    pub snippet: Option<VideoSnippet>,
    pub statistics: Option<VideoStatistics>,
    pub status: Option<VideoStatus>,
    pub suggestions: Option<VideoSuggestions>,
    pub topic_details: Option<VideoTopicDetails>,

}

A video resource represents a YouTube video.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete videos (none)

get rating videos (none)

insert videos (request|response)

list videos (none)

rate videos (none)

report abuse videos (none)

update videos (request|response)

Fields
age_gating: Option<VideoAgeGating>

Age restriction details related to a video. This data can only be retrieved by the video owner.

content_details: Option<VideoContentDetails>

The contentDetails object contains information about the video content, including the length of the video and its aspect ratio.

etag: Option<String>

Etag of this resource.

file_details: Option<VideoFileDetails>

The fileDetails object encapsulates information about the video file that was uploaded to YouTube, including the file’s resolution, duration, audio and video codecs, stream bitrates, and more. This data can only be retrieved by the video owner.

id: Option<String>

The ID that YouTube uses to uniquely identify the video.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string “youtube#video”.

live_streaming_details: Option<VideoLiveStreamingDetails>

The liveStreamingDetails object contains metadata about a live video broadcast. The object will only be present in a video resource if the video is an upcoming, live, or completed live broadcast.

localizations: Option<HashMap<String, VideoLocalization>>

The localizations object contains localized versions of the basic details about the video, such as its title and description.

monetization_details: Option<VideoMonetizationDetails>

The monetizationDetails object encapsulates information about the monetization status of the video.

player: Option<VideoPlayer>

The player object contains information that you would use to play the video in an embedded player.

processing_details: Option<VideoProcessingDetails>

The processingDetails object encapsulates information about YouTube’s progress in processing the uploaded video file. The properties in the object identify the current processing status and an estimate of the time remaining until YouTube finishes processing the video. This part also indicates whether different types of data or content, such as file details or thumbnail images, are available for the video. The processingProgress object is designed to be polled so that the video uploaded can track the progress that YouTube has made in processing the uploaded video file. This data can only be retrieved by the video owner.

project_details: Option<VideoProjectDetails>

The projectDetails object contains information about the project specific video metadata. b/157517979: This part was never populated after it was added. However, it sees non-zero traffic because there is generated client code in the wild that refers to it [1]. We keep this field and do NOT remove it because otherwise V3 would return an error when this part gets requested [2]. [1] https://developers.google.com/resources/api-libraries/documentation/youtube/v3/csharp/latest/classGoogle_1_1Apis_1_1YouTube_1_1v3_1_1Data_1_1VideoProjectDetails.html [2] http://google3/video/youtube/src/python/servers/data_api/common.py?l=1565-1569&rcl=344141677

recording_details: Option<VideoRecordingDetails>

The recordingDetails object encapsulates information about the location, date and address where the video was recorded.

snippet: Option<VideoSnippet>

The snippet object contains basic details about the video, such as its title, description, and category.

statistics: Option<VideoStatistics>

The statistics object contains statistics about the video.

status: Option<VideoStatus>

The status object contains information about the video’s uploading, processing, and privacy statuses.

suggestions: Option<VideoSuggestions>

The suggestions object encapsulates suggestions that identify opportunities to improve the video quality or the metadata for the uploaded video. This data can only be retrieved by the video owner.

topic_details: Option<VideoTopicDetails>

The topicDetails object encapsulates information about Freebase topics associated with the video.