Struct google_youtube3::api::VideoSuggestions

pub struct VideoSuggestions {
    pub editor_suggestions: Option<Vec<String>>,
    pub processing_errors: Option<Vec<String>>,
    pub processing_hints: Option<Vec<String>>,
    pub processing_warnings: Option<Vec<String>>,
    pub tag_suggestions: Option<Vec<VideoSuggestionsTagSuggestion>>,
}

Specifies suggestions on how to improve video content, including encoding hints, tag suggestions, and editor suggestions.

This type is not used in any activity, and only used as part of another schema.

Fields
editor_suggestions: Option<Vec<String>>

A list of video editing operations that might improve the video quality or playback experience of the uploaded video.

processing_errors: Option<Vec<String>>

A list of errors that will prevent YouTube from successfully processing the uploaded video video. These errors indicate that, regardless of the video’s current processing status, eventually, that status will almost certainly be failed.

processing_hints: Option<Vec<String>>

A list of suggestions that may improve YouTube’s ability to process the video.

processing_warnings: Option<Vec<String>>

A list of reasons why YouTube may have difficulty transcoding the uploaded video or that might result in an erroneous transcoding. These warnings are generated before YouTube actually processes the uploaded video file. In addition, they identify issues that are unlikely to cause the video processing to fail but that might cause problems such as sync issues, video artifacts, or a missing audio track.

tag_suggestions: Option<Vec<VideoSuggestionsTagSuggestion>>

A list of keyword tags that could be added to the video’s metadata to increase the likelihood that users will locate your video when searching or browsing on YouTube.