Struct google_youtube3::api::WatermarkMethods

pub struct WatermarkMethods<'a, S>
where
    S: 'a,
{ /* private fields */ }

A builder providing access to all methods supported on watermark resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use youtube3::{YouTube, oauth2, hyper, hyper_rustls, chrono, FieldMask};
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = oauth2::InstalledFlowAuthenticator::builder(
        secret,
        oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = YouTube::new(hyper::Client::builder().build(hyper_rustls::HttpsConnectorBuilder::new().with_native_roots().https_or_http().enable_http1().build()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `set(...)` and `unset(...)`
// to build up your call.
let rb = hub.watermarks();

Implementations

impl<'a, S> WatermarkMethods<'a, S>

pub fn set( &self, request: InvideoBranding, channel_id: &str ) -> WatermarkSetCall<'a, S>

Create a builder to help you perform the following task:

Allows upload of watermark image and setting it for a channel.

Arguments

request - No description provided.

channelId - No description provided.

pub fn unset(&self, channel_id: &str) -> WatermarkUnsetCall<'a, S>

Create a builder to help you perform the following task:

Allows removal of channel watermark.

Arguments

channelId - No description provided.