Struct google_youtube3::ActivityContentDetailsPromotedItem

pub struct ActivityContentDetailsPromotedItem {
    pub cta_type: Option<String>,
    pub ad_tag: Option<String>,
    pub destination_url: Option<String>,
    pub forecasting_url: Option<Vec<String>>,
    pub impression_url: Option<Vec<String>>,
    pub creative_view_url: Option<String>,
    pub video_id: Option<String>,
    pub description_text: Option<String>,
    pub custom_cta_button_text: Option<String>,
    pub click_tracking_url: Option<String>,
}

Details about a resource which is being promoted.

This type is not used in any activity, and only used as part of another schema.

Fields
cta_type: Option<String>

The type of call-to-action, a message to the user indicating action that can be taken.

ad_tag: Option<String>

The URL the client should fetch to request a promoted item.

destination_url: Option<String>

The URL the client should direct the user to, if the user chooses to visit the advertiser's website.

forecasting_url: Option<Vec<String>>

The list of forecasting URLs. The client should ping all of these URLs when a promoted item is not available, to indicate that a promoted item could have been shown.

impression_url: Option<Vec<String>>

The list of impression URLs. The client should ping all of these URLs to indicate that the user was shown this promoted item.

creative_view_url: Option<String>

The URL the client should ping to indicate that the user was shown this promoted item.

video_id: Option<String>

The ID that YouTube uses to uniquely identify the promoted video.

description_text: Option<String>

The text description to accompany the promoted item.

custom_cta_button_text: Option<String>

The custom call-to-action button text. If specified, it will override the default button text for the cta_type.

click_tracking_url: Option<String>

The URL the client should ping to indicate that the user clicked through on this promoted item.