Struct google_youtube3::ActivityListCall

pub struct ActivityListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Returns a list of channel activity events that match the request criteria. For example, you can retrieve events associated with a particular channel, events associated with the user's subscriptions and Google+ friends, or the YouTube home page feed, which is customized for each user.

A builder for the list method supported by a activity resource. It is not used directly, but through a ActivityMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.activities().list("part")
             .region_code("vero")
             .published_before("diam")
             .published_after("rebum.")
             .page_token("consetetur")
             .mine(true)
             .max_results(25)
             .home(false)
             .channel_id("invidunt")
             .doit();

Methods

impl<'a, C, A> ActivityListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ActivityListResponse)>

Perform the operation you have build so far.

fn part(self, new_value: &str) -> ActivityListCall<'a, C, A>

The part parameter specifies a comma-separated list of one or more activity resource properties that the API response will include.

If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in an activity resource, the snippet property contains other properties that identify the type of activity, a display title for the activity, and so forth. If you set part=snippet, the API response will also contain all of those nested properties.

Sets the part query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn region_code(self, new_value: &str) -> ActivityListCall<'a, C, A>

The regionCode parameter instructs the API to return results for the specified country. The parameter value is an ISO 3166-1 alpha-2 country code. YouTube uses this value when the authorized user's previous activity on YouTube does not provide enough information to generate the activity feed.

Sets the region code query property to the given value.

fn published_before(self, new_value: &str) -> ActivityListCall<'a, C, A>

The publishedBefore parameter specifies the date and time before which an activity must have occurred for that activity to be included in the API response. If the parameter value specifies a day, but not a time, then any activities that occurred that day will be excluded from the result set. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

Sets the published before query property to the given value.

fn published_after(self, new_value: &str) -> ActivityListCall<'a, C, A>

The publishedAfter parameter specifies the earliest date and time that an activity could have occurred for that activity to be included in the API response. If the parameter value specifies a day, but not a time, then any activities that occurred that day will be included in the result set. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

Sets the published after query property to the given value.

fn page_token(self, new_value: &str) -> ActivityListCall<'a, C, A>

The pageToken parameter identifies a specific page in the result set that should be returned. In an API response, the nextPageToken and prevPageToken properties identify other pages that could be retrieved.

Sets the page token query property to the given value.

fn mine(self, new_value: bool) -> ActivityListCall<'a, C, A>

Set this parameter's value to true to retrieve a feed of the authenticated user's activities.

Sets the mine query property to the given value.

fn max_results(self, new_value: u32) -> ActivityListCall<'a, C, A>

The maxResults parameter specifies the maximum number of items that should be returned in the result set.

Sets the max results query property to the given value.

fn home(self, new_value: bool) -> ActivityListCall<'a, C, A>

Set this parameter's value to true to retrieve the activity feed that displays on the YouTube home page for the currently authenticated user.

Sets the home query property to the given value.

fn channel_id(self, new_value: &str) -> ActivityListCall<'a, C, A>

The channelId parameter specifies a unique YouTube channel ID. The API will then return a list of that channel's activities.

Sets the channel id query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> ActivityListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ActivityListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> ActivityListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Readonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.