Struct google_youtube3::ActivityMethods

pub struct ActivityMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on activity resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `insert(...)` and `list(...)`
// to build up your call.
let rb = hub.activities();

Methods

impl<'a, C, A> ActivityMethods<'a, C, A>

fn list(&self, part: &str) -> ActivityListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of channel activity events that match the request criteria. For example, you can retrieve events associated with a particular channel, events associated with the user's subscriptions and Google+ friends, or the YouTube home page feed, which is customized for each user.

Arguments

part - The part parameter specifies a comma-separated list of one or more activity resource properties that the API response will include. If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in an activity resource, the snippet property contains other properties that identify the type of activity, a display title for the activity, and so forth. If you set part=snippet, the API response will also contain all of those nested properties.

fn insert(&self, request: Activity) -> ActivityInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Posts a bulletin for a specific channel. (The user submitting the request must be authorized to act on the channel's behalf.)

Note: Even though an activity resource can contain information about actions like a user rating a video or marking a video as a favorite, you need to use other API methods to generate those activity resources. For example, you would use the API's videos.rate() method to rate a video and the playlistItems.insert() method to mark a video as a favorite.

Arguments

request - No description provided.