Struct google_youtube3::CaptionSnippet

pub struct CaptionSnippet {
    pub status: Option<String>,
    pub is_cc: Option<bool>,
    pub audio_track_type: Option<String>,
    pub language: Option<String>,
    pub video_id: Option<String>,
    pub is_draft: Option<bool>,
    pub is_auto_synced: Option<bool>,
    pub last_updated: Option<String>,
    pub track_kind: Option<String>,
    pub is_easy_reader: Option<bool>,
    pub is_large: Option<bool>,
    pub failure_reason: Option<String>,
    pub name: Option<String>,
}

Basic details about a caption track, such as its language and name.

This type is not used in any activity, and only used as part of another schema.

Fields

status

The caption track's status.

is_cc

Indicates whether the track contains closed captions for the deaf and hard of hearing. The default value is false.

audio_track_type

The type of audio track associated with the caption track.

language

The language of the caption track. The property value is a BCP-47 language tag.

video_id

The ID that YouTube uses to uniquely identify the video associated with the caption track.

is_draft

Indicates whether the caption track is a draft. If the value is true, then the track is not publicly visible. The default value is false.

is_auto_synced

Indicates whether YouTube synchronized the caption track to the audio track in the video. The value will be true if a sync was explicitly requested when the caption track was uploaded. For example, when calling the captions.insert or captions.update methods, you can set the sync parameter to true to instruct YouTube to sync the uploaded track to the video. If the value is false, YouTube uses the time codes in the uploaded caption track to determine when to display captions.

last_updated

The date and time when the caption track was last updated. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

track_kind

The caption track's type.

is_easy_reader

Indicates whether caption track is formatted for "easy reader," meaning it is at a third-grade level for language learners. The default value is false.

is_large

Indicates whether the caption track uses large text for the vision-impaired. The default value is false.

failure_reason

The reason that YouTube failed to process the caption track. This property is only present if the state property's value is failed.

name

The name of the caption track. The name is intended to be visible to the user as an option during playback.