Struct google_youtube3::ChannelAuditDetails

pub struct ChannelAuditDetails {
    pub copyright_strikes_good_standing: Option<bool>,
    pub community_guidelines_good_standing: Option<bool>,
    pub content_id_claims_good_standing: Option<bool>,
    pub overall_good_standing: Option<bool>,
}

The auditDetails object encapsulates channel data that is relevant for YouTube Partners during the audit process.

This type is not used in any activity, and only used as part of another schema.

Fields
copyright_strikes_good_standing: Option<bool>

Whether or not the channel has any copyright strikes.

community_guidelines_good_standing: Option<bool>

Whether or not the channel respects the community guidelines.

content_id_claims_good_standing: Option<bool>

Whether or not the channel has any unresolved claims.

overall_good_standing: Option<bool>

Describes the general state of the channel. This field will always show if there are any issues whatsoever with the channel. Currently this field represents the result of the logical and operation over the community guidelines good standing, the copyright strikes good standing and the content ID claims good standing, but this may change in the future.