Struct google_youtube3::ChannelBannerInsertCall

pub struct ChannelBannerInsertCall<'a, C, A> where C: 'a, A: 'a { /* fields omitted */ }

Uploads a channel banner image to YouTube. This method represents the first two steps in a three-step process to update the banner image for a channel:

Call the channelBanners.insert method to upload the binary image data to YouTube. The image must have a 16:9 aspect ratio and be at least 2120x1192 pixels.

Extract the url property's value from the response that the API returns for step 1.

Call the channels.update method to update the channel's branding settings. Set the brandingSettings.image.bannerExternalUrl property's value to the URL obtained in step 2.

A builder for the insert method supported by a channelBanner resource. It is not used directly, but through a ChannelBannerMethods instance.

Example

Instantiate a resource method builder

use youtube3::ChannelBannerResource;
use std::fs;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = ChannelBannerResource::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload(...)`.
// Values shown here are possibly random and not representative !
let result = hub.channel_banners().insert(req)
             .on_behalf_of_content_owner("consetetur")
             .upload(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());

Methods

impl<'a, C, A> ChannelBannerInsertCall<'a, C, A> where C: BorrowMut<Client>,
A: GetToken

fn upload<RS>(self,
stream: RS,
mime_type: Mime)
-> Result<(Response, ChannelBannerResource)> where RS: ReadSeek

Upload media all at once. If the upload fails for whichever reason, all progress is lost.

max size: 6MB

multipart: yes

valid mime types: 'application/octet-stream', 'image/jpeg' and 'image/png'

fn upload_resumable<RS>(self,
resumeable_stream: RS,
mime_type: Mime)
-> Result<(Response, ChannelBannerResource)> where RS: ReadSeek

Upload media in a resumable fashion. Even if the upload fails or is interrupted, it can be resumed for a certain amount of time as the server maintains state temporarily.

The delegate will be asked for an upload_url(), and if not provided, will be asked to store an upload URL that was provided by the server, using store_upload_url(...). The upload will be done in chunks, the delegate may specify the chunk_size() and may cancel the operation before each chunk is uploaded, using cancel_chunk_upload(...).

max size: 6MB

multipart: yes

valid mime types: 'application/octet-stream', 'image/jpeg' and 'image/png'

fn request(self,
new_value: ChannelBannerResource)
-> ChannelBannerInsertCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn on_behalf_of_content_owner(self,
new_value: &str)
-> ChannelBannerInsertCall<'a, C, A>

Note: This parameter is intended exclusively for YouTube content partners.

The onBehalfOfContentOwner parameter indicates that the request's authorization credentials identify a YouTube CMS user who is acting on behalf of the content owner specified in the parameter value. This parameter is intended for YouTube content partners that own and manage many different YouTube channels. It allows content owners to authenticate once and get access to all their video and channel data, without having to provide authentication credentials for each individual channel. The CMS account that the user authenticates with must be linked to the specified YouTube content owner.

Sets the on behalf of content owner query property to the given value.

fn delegate(self,
new_value: &'a mut Delegate)
-> ChannelBannerInsertCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ChannelBannerInsertCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> ChannelBannerInsertCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.