Struct google_youtube3::ChannelBannerMethods

pub struct ChannelBannerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on channelBanner resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `insert(...)`
// to build up your call.
let rb = hub.channel_banners();

Methods

impl<'a, C, A> ChannelBannerMethods<'a, C, A>

fn insert(&self, request: ChannelBannerResource) -> ChannelBannerInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads a channel banner image to YouTube. This method represents the first two steps in a three-step process to update the banner image for a channel:

Call the channelBanners.insert method to upload the binary image data to YouTube. The image must have a 16:9 aspect ratio and be at least 2120x1192 pixels.

Extract the url property's value from the response that the API returns for step 1.

Call the channels.update method to update the channel's branding settings. Set the brandingSettings.image.bannerExternalUrl property's value to the URL obtained in step 2.

Arguments

request - No description provided.