Struct google_youtube3::ChannelListResponse

pub struct ChannelListResponse {
    pub event_id: Option<String>,
    pub next_page_token: Option<String>,
    pub kind: Option<String>,
    pub visitor_id: Option<String>,
    pub items: Option<Vec<Channel>>,
    pub token_pagination: Option<TokenPagination>,
    pub etag: Option<String>,
    pub prev_page_token: Option<String>,
    pub page_info: Option<PageInfo>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list channels (response)

Fields

event_id

Serialized EventId of the request which produced this response.

next_page_token

The token that can be used as the value of the pageToken parameter to retrieve the next page in the result set.

kind

Identifies what kind of resource this is. Value: the fixed string "youtube#channelListResponse".

visitor_id

The visitorId identifies the visitor.

items

A list of channels that match the request criteria.

token_pagination

no description provided

etag

Etag of this resource.

prev_page_token

The token that can be used as the value of the pageToken parameter to retrieve the previous page in the result set.

page_info

no description provided