Struct google_youtube3::ChannelMethods

pub struct ChannelMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on channel resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)` and `update(...)`
// to build up your call.
let rb = hub.channels();

Methods

impl<'a, C, A> ChannelMethods<'a, C, A>

fn update(&self, request: Channel) -> ChannelUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a channel's metadata. Note that this method currently only supports updates to the channel resource's brandingSettings and invideoPromotion objects and their child properties.

Arguments

request - No description provided.

fn list(&self, part: &str) -> ChannelListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a collection of zero or more channel resources that match the request criteria.

Arguments

part - The part parameter specifies a comma-separated list of one or more channel resource properties that the API response will include. If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in a channel resource, the contentDetails property contains other properties, such as the uploads properties. As such, if you set part=contentDetails, the API response will also contain all of those nested properties.