Struct google_youtube3::ChannelSectionMethods

pub struct ChannelSectionMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on channelSection resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `insert(...)`, `list(...)` and `update(...)`
// to build up your call.
let rb = hub.channel_sections();

Methods

impl<'a, C, A> ChannelSectionMethods<'a, C, A>

fn list(&self, part: &str) -> ChannelSectionListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns channelSection resources that match the API request criteria.

Arguments

part - The part parameter specifies a comma-separated list of one or more channelSection resource properties that the API response will include. The part names that you can include in the parameter value are id, snippet, and contentDetails. If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in a channelSection resource, the snippet property contains other properties, such as a display title for the channelSection. If you set part=snippet, the API response will also contain all of those nested properties.

fn insert(&self, request: ChannelSection) -> ChannelSectionInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a channelSection for the authenticated user's channel.

Arguments

request - No description provided.

fn delete(&self, id: &str) -> ChannelSectionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a channelSection.

Arguments

id - The id parameter specifies the YouTube channelSection ID for the resource that is being deleted. In a channelSection resource, the id property specifies the YouTube channelSection ID.

fn update(&self, request: ChannelSection) -> ChannelSectionUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update a channelSection.

Arguments

request - No description provided.