Struct google_youtube3::ChannelSectionSnippet

pub struct ChannelSectionSnippet {
    pub style: Option<String>,
    pub localized: Option<ChannelSectionLocalization>,
    pub title: Option<String>,
    pub position: Option<u32>,
    pub channel_id: Option<String>,
    pub type_: Option<String>,
    pub default_language: Option<String>,
}

Basic details about a channel section, including title, style and position.

This type is not used in any activity, and only used as part of another schema.

Fields
style: Option<String>

The style of the channel section.

localized: Option<ChannelSectionLocalization>

Localized title, read-only.

title: Option<String>

The channel section's title for multiple_playlists and multiple_channels.

position: Option<u32>

The position of the channel section in the channel.

channel_id: Option<String>

The ID that YouTube uses to uniquely identify the channel that published the channel section.

type_: Option<String>

The type of the channel section.

default_language: Option<String>

The language of the channel section's default title and description.