Struct google_youtube3::ChannelSettings

pub struct ChannelSettings {
    pub description: Option<String>,
    pub title: Option<String>,
    pub country: Option<String>,
    pub show_browse_view: Option<bool>,
    pub featured_channels_title: Option<String>,
    pub default_language: Option<String>,
    pub unsubscribed_trailer: Option<String>,
    pub featured_channels_urls: Option<Vec<String>>,
    pub profile_color: Option<String>,
    pub default_tab: Option<String>,
    pub moderate_comments: Option<bool>,
    pub keywords: Option<String>,
    pub show_related_channels: Option<bool>,
    pub tracking_analytics_account_id: Option<String>,
}

Branding properties for the channel view.

This type is not used in any activity, and only used as part of another schema.

Fields

description

Specifies the channel description.

title

Specifies the channel title.

country

The country of the channel.

show_browse_view

Whether the tab to browse the videos should be displayed.

featured_channels_title

Title for the featured channels tab.

default_language

no description provided

unsubscribed_trailer

The trailer of the channel, for users that are not subscribers.

featured_channels_urls

The list of featured channels.

profile_color

A prominent color that can be rendered on this channel page.

default_tab

Which content tab users should see when viewing the channel.

moderate_comments

Whether user-submitted comments left on the channel page need to be approved by the channel owner to be publicly visible.

keywords

Lists keywords associated with the channel, comma-separated.

show_related_channels

Whether related channels should be proposed.

tracking_analytics_account_id

The ID for a Google Analytics account to track and measure traffic to the channels.