Struct google_youtube3::CommentMethods

pub struct CommentMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on comment resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `insert(...)`, `list(...)`, `mark_as_spam(...)`, `set_moderation_status(...)` and `update(...)`
// to build up your call.
let rb = hub.comments();

Methods

impl<'a, C, A> CommentMethods<'a, C, A>

fn insert(&self, request: Comment) -> CommentInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a reply to an existing comment. Note: To create a top-level comment, use the commentThreads.insert method.

Arguments

request - No description provided.

fn set_moderation_status(&self, id: &str, moderation_status: &str) -> CommentSetModerationStatuCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the moderation status of one or more comments. The API request must be authorized by the owner of the channel or video associated with the comments.

Arguments

id - The id parameter specifies a comma-separated list of IDs that identify the comments for which you are updating the moderation status.

moderationStatus - Identifies the new moderation status of the specified comments.

fn delete(&self, id: &str) -> CommentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a comment.

Arguments

id - The id parameter specifies the comment ID for the resource that is being deleted.

fn list(&self, part: &str) -> CommentListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of comments that match the API request parameters.

Arguments

part - The part parameter specifies a comma-separated list of one or more comment resource properties that the API response will include.

fn update(&self, request: Comment) -> CommentUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies a comment.

Arguments

request - No description provided.

fn mark_as_spam(&self, id: &str) -> CommentMarkAsSpamCall<'a, C, A>

Create a builder to help you perform the following task:

Expresses the caller's opinion that one or more comments should be flagged as spam.

Arguments

id - The id parameter specifies a comma-separated list of IDs of comments that the caller believes should be classified as spam.