Struct google_youtube3::CommentThread

pub struct CommentThread {
    pub snippet: Option<CommentThreadSnippet>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub id: Option<String>,
    pub replies: Option<CommentThreadReplies>,
}

A comment thread represents information that applies to a top level comment and all its replies. It can also include the top level comment itself and some of the replies.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update comment threads (request|response)

list comment threads (none)

insert comment threads (request|response)

Fields
snippet: Option<CommentThreadSnippet>

The snippet object contains basic details about the comment thread and also the top level comment.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "youtube#commentThread".

etag: Option<String>

Etag of this resource.

id: Option<String>

The ID that YouTube uses to uniquely identify the comment thread.

replies: Option<CommentThreadReplies>

The replies object contains a limited number of replies (if any) to the top level comment found in the snippet.