Struct google_youtube3::CommentThreadListCall

pub struct CommentThreadListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Returns a list of comment threads that match the API request parameters.

A builder for the list method supported by a commentThread resource. It is not used directly, but through a CommentThreadMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.comment_threads().list("part")
             .video_id("kasd")
             .text_format("invidunt")
             .search_terms("rebum.")
             .page_token("Lorem")
             .order("clita")
             .moderation_status("invidunt")
             .max_results(11)
             .id("At")
             .channel_id("consetetur")
             .all_threads_related_to_channel_id("et")
             .doit();

Methods

impl<'a, C, A> CommentThreadListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, CommentThreadListResponse)>

Perform the operation you have build so far.

fn part(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The part parameter specifies a comma-separated list of one or more commentThread resource properties that the API response will include.

Sets the part query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn video_id(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The videoId parameter instructs the API to return comment threads associated with the specified video ID.

Sets the video id query property to the given value.

fn text_format(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

Set this parameter's value to html or plainText to instruct the API to return the comments left by users in html formatted or in plain text.

Sets the text format query property to the given value.

fn search_terms(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The searchTerms parameter instructs the API to limit the API response to only contain comments that contain the specified search terms.

Note: This parameter is not supported for use in conjunction with the id parameter.

Sets the search terms query property to the given value.

fn page_token(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The pageToken parameter identifies a specific page in the result set that should be returned. In an API response, the nextPageToken property identifies the next page of the result that can be retrieved.

Note: This parameter is not supported for use in conjunction with the id parameter.

Sets the page token query property to the given value.

fn order(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The order parameter specifies the order in which the API response should list comment threads. Valid values are: - time - Comment threads are ordered by time. This is the default behavior. - relevance - Comment threads are ordered by relevance.Note: This parameter is not supported for use in conjunction with the id parameter.

Sets the order query property to the given value.

fn moderation_status(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

Set this parameter to limit the returned comment threads to a particular moderation state.

Note: This parameter is not supported for use in conjunction with the id parameter.

Sets the moderation status query property to the given value.

fn max_results(self, new_value: u32) -> CommentThreadListCall<'a, C, A>

The maxResults parameter specifies the maximum number of items that should be returned in the result set.

Note: This parameter is not supported for use in conjunction with the id parameter.

Sets the max results query property to the given value.

fn id(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The id parameter specifies a comma-separated list of comment thread IDs for the resources that should be retrieved.

Sets the id query property to the given value.

fn channel_id(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The channelId parameter instructs the API to return comment threads containing comments about the specified channel. (The response will not include comments left on videos that the channel uploaded.)

Sets the channel id query property to the given value.

fn all_threads_related_to_channel_id(self, new_value: &str) -> CommentThreadListCall<'a, C, A>

The allThreadsRelatedToChannelId parameter instructs the API to return all comment threads associated with the specified channel. The response can include comments about the channel or about the channel's videos.

Sets the all threads related to channel id query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> CommentThreadListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> CommentThreadListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> CommentThreadListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::ForceSsl.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.