Struct google_youtube3::I18nLanguageListResponse

pub struct I18nLanguageListResponse {
    pub event_id: Option<String>,
    pub items: Option<Vec<I18nLanguage>>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub visitor_id: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list i18n languages (response)

Fields

event_id

Serialized EventId of the request which produced this response.

items

A list of supported i18n languages. In this map, the i18n language ID is the map key, and its value is the corresponding i18nLanguage resource.

kind

Identifies what kind of resource this is. Value: the fixed string "youtube#i18nLanguageListResponse".

etag

Etag of this resource.

visitor_id

The visitorId identifies the visitor.