Struct google_youtube3::ImageSettings

pub struct ImageSettings {
    pub banner_tv_medium_image_url: Option<String>,
    pub large_branded_banner_image_imap_script: Option<LocalizedProperty>,
    pub banner_mobile_image_url: Option<String>,
    pub small_branded_banner_image_url: Option<LocalizedProperty>,
    pub banner_tablet_hd_image_url: Option<String>,
    pub banner_tablet_low_image_url: Option<String>,
    pub banner_mobile_medium_hd_image_url: Option<String>,
    pub tracking_image_url: Option<String>,
    pub banner_mobile_extra_hd_image_url: Option<String>,
    pub banner_tablet_image_url: Option<String>,
    pub banner_mobile_low_image_url: Option<String>,
    pub banner_tv_image_url: Option<String>,
    pub banner_tv_low_image_url: Option<String>,
    pub banner_tablet_extra_hd_image_url: Option<String>,
    pub large_branded_banner_image_url: Option<LocalizedProperty>,
    pub banner_tv_high_image_url: Option<String>,
    pub background_image_url: Option<LocalizedProperty>,
    pub small_branded_banner_image_imap_script: Option<LocalizedProperty>,
    pub banner_image_url: Option<String>,
    pub banner_mobile_hd_image_url: Option<String>,
    pub banner_external_url: Option<String>,
    pub watch_icon_image_url: Option<String>,
}

Branding properties for images associated with the channel.

This type is not used in any activity, and only used as part of another schema.

Fields

banner_tv_medium_image_url

Banner image. TV size medium resolution (1280x720).

large_branded_banner_image_imap_script

The image map script for the large banner image.

banner_mobile_image_url

Banner image. Mobile size (640x175).

small_branded_banner_image_url

The URL for the 640px by 70px banner image that appears below the video player in the default view of the video watch page.

banner_tablet_hd_image_url

Banner image. Tablet size high resolution (2276x377).

banner_tablet_low_image_url

Banner image. Tablet size low resolution (1138x188).

banner_mobile_medium_hd_image_url

Banner image. Mobile size medium/high resolution (960x263).

tracking_image_url

The URL for a 1px by 1px tracking pixel that can be used to collect statistics for views of the channel or video pages.

banner_mobile_extra_hd_image_url

Banner image. Mobile size high resolution (1440x395).

banner_tablet_image_url

Banner image. Tablet size (1707x283).

banner_mobile_low_image_url

Banner image. Mobile size low resolution (320x88).

banner_tv_image_url

Banner image. TV size extra high resolution (2120x1192).

banner_tv_low_image_url

Banner image. TV size low resolution (854x480).

banner_tablet_extra_hd_image_url

Banner image. Tablet size extra high resolution (2560x424).

large_branded_banner_image_url

The URL for the 854px by 70px image that appears below the video player in the expanded video view of the video watch page.

banner_tv_high_image_url

Banner image. TV size high resolution (1920x1080).

background_image_url

The URL for the background image shown on the video watch page. The image should be 1200px by 615px, with a maximum file size of 128k.

small_branded_banner_image_imap_script

The image map script for the small banner image.

banner_image_url

Banner image. Desktop size (1060x175).

banner_mobile_hd_image_url

Banner image. Mobile size high resolution (1280x360).

banner_external_url

This is used only in update requests; if it's set, we use this URL to generate all of the above banner URLs.

watch_icon_image_url

The URL for the image that appears above the top-left corner of the video player. This is a 25-pixel-high image with a flexible width that cannot exceed 170 pixels.