Struct google_youtube3::IngestionInfo

pub struct IngestionInfo {
    pub backup_ingestion_address: Option<String>,
    pub stream_name: Option<String>,
    pub ingestion_address: Option<String>,
}

Describes information necessary for ingesting an RTMP or an HTTP stream.

This type is not used in any activity, and only used as part of another schema.

Fields

backup_ingestion_address

The backup ingestion URL that you should use to stream video to YouTube. You have the option of simultaneously streaming the content that you are sending to the ingestionAddress to this URL.

stream_name

The HTTP or RTMP stream name that YouTube assigns to the video stream.

ingestion_address

The primary ingestion URL that you should use to stream video to YouTube. You must stream video to this URL.

Depending on which application or tool you use to encode your video stream, you may need to enter the stream URL and stream name separately or you may need to concatenate them in the following format:

STREAM_URL/STREAM_NAME