Struct google_youtube3::LiveBroadcastContentDetails

pub struct LiveBroadcastContentDetails {
    pub start_with_slate: Option<bool>,
    pub bound_stream_id: Option<String>,
    pub enable_embed: Option<bool>,
    pub enable_closed_captions: Option<bool>,
    pub enable_content_encryption: Option<bool>,
    pub record_from_start: Option<bool>,
    pub enable_dvr: Option<bool>,
    pub monitor_stream: Option<MonitorStreamInfo>,
}

Detailed settings of a broadcast.

This type is not used in any activity, and only used as part of another schema.

Fields

start_with_slate

This setting indicates whether the broadcast should automatically begin with an in-stream slate when you update the broadcast's status to live. After updating the status, you then need to send a liveCuepoints.insert request that sets the cuepoint's eventState to end to remove the in-stream slate and make your broadcast stream visible to viewers.

bound_stream_id

This value uniquely identifies the live stream bound to the broadcast.

enable_embed

This setting indicates whether the broadcast video can be played in an embedded player. If you choose to archive the video (using the enableArchive property), this setting will also apply to the archived video.

enable_closed_captions

This setting indicates whether closed captioning is enabled for this broadcast. The ingestion URL of the closed captions is returned through the liveStreams API.

enable_content_encryption

This setting indicates whether YouTube should enable content encryption for the broadcast.

record_from_start

Automatically start recording after the event goes live. The default value for this property is true.

Important: You must also set the enableDvr property's value to true if you want the playback to be available immediately after the broadcast ends. If you set this property's value to true but do not also set the enableDvr property to true, there may be a delay of around one day before the archived video will be available for playback.

enable_dvr

This setting determines whether viewers can access DVR controls while watching the video. DVR controls enable the viewer to control the video playback experience by pausing, rewinding, or fast forwarding content. The default value for this property is true.

Important: You must set the value to true and also set the enableArchive property's value to true if you want to make playback available immediately after the broadcast ends.

monitor_stream

The monitorStream object contains information about the monitor stream, which the broadcaster can use to review the event content before the broadcast stream is shown publicly.