Struct google_youtube3::LiveBroadcastSnippet

pub struct LiveBroadcastSnippet {
    pub actual_end_time: Option<String>,
    pub description: Option<String>,
    pub title: Option<String>,
    pub channel_id: Option<String>,
    pub published_at: Option<String>,
    pub scheduled_start_time: Option<String>,
    pub actual_start_time: Option<String>,
    pub scheduled_end_time: Option<String>,
    pub thumbnails: Option<ThumbnailDetails>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

actual_end_time

The date and time that the broadcast actually ended. This information is only available once the broadcast's state is complete. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

description

The broadcast's description. As with the title, you can set this field by modifying the broadcast resource or by setting the description field of the corresponding video resource.

title

The broadcast's title. Note that the broadcast represents exactly one YouTube video. You can set this field by modifying the broadcast resource or by setting the title field of the corresponding video resource.

channel_id

The ID that YouTube uses to uniquely identify the channel that is publishing the broadcast.

published_at

The date and time that the broadcast was added to YouTube's live broadcast schedule. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

scheduled_start_time

The date and time that the broadcast is scheduled to start. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

actual_start_time

The date and time that the broadcast actually started. This information is only available once the broadcast's state is live. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

scheduled_end_time

The date and time that the broadcast is scheduled to end. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

thumbnails

A map of thumbnail images associated with the broadcast. For each nested object in this object, the key is the name of the thumbnail image, and the value is an object that contains other information about the thumbnail.