Struct google_youtube3::LiveBroadcastStatus

pub struct LiveBroadcastStatus {
    pub recording_status: Option<String>,
    pub privacy_status: Option<String>,
    pub life_cycle_status: Option<String>,
    pub live_broadcast_priority: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

recording_status

The broadcast's recording status.

privacy_status

The broadcast's privacy status. Note that the broadcast represents exactly one YouTube video, so the privacy settings are identical to those supported for videos. In addition, you can set this field by modifying the broadcast resource or by setting the privacyStatus field of the corresponding video resource.

life_cycle_status

The broadcast's status. The status can be updated using the API's liveBroadcasts.transition method.

live_broadcast_priority

Priority of the live broadcast event (internal state).