Struct google_youtube3::LiveChatMessageSnippet

pub struct LiveChatMessageSnippet {
    pub message_retracted_details: Option<LiveChatMessageRetractedDetails>,
    pub has_display_content: Option<bool>,
    pub published_at: Option<String>,
    pub live_chat_id: Option<String>,
    pub poll_edited_details: Option<LiveChatPollEditedDetails>,
    pub text_message_details: Option<LiveChatTextMessageDetails>,
    pub poll_voted_details: Option<LiveChatPollVotedDetails>,
    pub display_message: Option<String>,
    pub super_chat_details: Option<LiveChatSuperChatDetails>,
    pub fan_funding_event_details: Option<LiveChatFanFundingEventDetails>,
    pub user_banned_details: Option<LiveChatUserBannedMessageDetails>,
    pub author_channel_id: Option<String>,
    pub message_deleted_details: Option<LiveChatMessageDeletedDetails>,
    pub poll_opened_details: Option<LiveChatPollOpenedDetails>,
    pub type_: Option<String>,
    pub poll_closed_details: Option<LiveChatPollClosedDetails>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
message_retracted_details: Option<LiveChatMessageRetractedDetails>

no description provided

has_display_content: Option<bool>

Whether the message has display content that should be displayed to users.

published_at: Option<String>

The date and time when the message was orignally published. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

live_chat_id: Option<String>

no description provided

poll_edited_details: Option<LiveChatPollEditedDetails>

no description provided

text_message_details: Option<LiveChatTextMessageDetails>

Details about the text message, this is only set if the type is 'textMessageEvent'.

poll_voted_details: Option<LiveChatPollVotedDetails>

no description provided

display_message: Option<String>

Contains a string that can be displayed to the user. If this field is not present the message is silent, at the moment only messages of type TOMBSTONE and CHAT_ENDED_EVENT are silent.

super_chat_details: Option<LiveChatSuperChatDetails>

Details about the Super Chat event, this is only set if the type is 'superChatEvent'.

fan_funding_event_details: Option<LiveChatFanFundingEventDetails>

Details about the funding event, this is only set if the type is 'fanFundingEvent'.

user_banned_details: Option<LiveChatUserBannedMessageDetails>

no description provided

author_channel_id: Option<String>

The ID of the user that authored this message, this field is not always filled. textMessageEvent - the user that wrote the message fanFundingEvent - the user that funded the broadcast newSponsorEvent - the user that just became a sponsor messageDeletedEvent - the moderator that took the action messageRetractedEvent - the author that retracted their message userBannedEvent - the moderator that took the action superChatEvent - the user that made the purchase

message_deleted_details: Option<LiveChatMessageDeletedDetails>

no description provided

poll_opened_details: Option<LiveChatPollOpenedDetails>

no description provided

type_: Option<String>

The type of message, this will always be present, it determines the contents of the message as well as which fields will be present.

poll_closed_details: Option<LiveChatPollClosedDetails>

no description provided