Struct google_youtube3::LiveStream

pub struct LiveStream {
    pub status: Option<LiveStreamStatus>,
    pub snippet: Option<LiveStreamSnippet>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub content_details: Option<LiveStreamContentDetails>,
    pub cdn: Option<CdnSettings>,
    pub id: Option<String>,
}

A live stream describes a live ingestion point.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete live streams (none)

update live streams (request|response)

list live streams (none)

insert live streams (request|response)

Fields

status

The status object contains information about live stream's status.

snippet

The snippet object contains basic details about the stream, including its channel, title, and description.

kind

Identifies what kind of resource this is. Value: the fixed string "youtube#liveStream".

etag

Etag of this resource.

content_details

The content_details object contains information about the stream, including the closed captions ingestion URL.

cdn

The cdn object defines the live stream's content delivery network (CDN) settings. These settings provide details about the manner in which you stream your content to YouTube.

id

The ID that YouTube assigns to uniquely identify the stream.