Struct google_youtube3::LiveStreamInsertCall

pub struct LiveStreamInsertCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Creates a video stream. The stream enables you to send your video to YouTube, which can then broadcast the video to your audience.

A builder for the insert method supported by a liveStream resource. It is not used directly, but through a LiveStreamMethods instance.

Settable Parts

id

snippet

cdn

status

Scopes

You will need authorization for at least one of the following scopes to make a valid call, possibly depending on parts:

https://www.googleapis.com/auth/youtube

https://www.googleapis.com/auth/youtube.force-ssl

The default scope will be Scope::Full.

Example

Instantiate a resource method builder

use youtube3::LiveStream;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = LiveStream::default();
req.status = Default::default(); // is LiveStreamStatus
req.snippet = Default::default(); // is LiveStreamSnippet
req.cdn = Default::default(); // is CdnSettings
req.id = Some("At".to_string());
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.live_streams().insert(req)
             .on_behalf_of_content_owner_channel("labore")
             .on_behalf_of_content_owner("invidunt")
             .doit();

Methods

impl<'a, C, A> LiveStreamInsertCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, LiveStream)>

Perform the operation you have build so far.

fn request(self, new_value: LiveStream) -> LiveStreamInsertCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

Settable Parts

id

snippet

cdn

status

fn part(self, new_value: &str) -> LiveStreamInsertCall<'a, C, A>

The part parameter serves two purposes in this operation. It identifies the properties that the write operation will set as well as the properties that the API response will include.

The part properties that you can include in the parameter value are id, snippet, cdn, and status.

Sets the part query property to the given value.

Even though the parts list is automatically derived from Resource passed in during instantiation and indicates which values you are passing, the response would contain the very same parts. This may not always be desirable, as you can obtain (newly generated) parts you cannot pass in, like statistics that are generated server side. Therefore you should use this method to specify the parts you provide in addition to the ones you want in the response.

Settable Parts

id

snippet

cdn

status

fn on_behalf_of_content_owner_channel(self, new_value: &str) -> LiveStreamInsertCall<'a, C, A>

This parameter can only be used in a properly authorized request. Note: This parameter is intended exclusively for YouTube content partners.

The onBehalfOfContentOwnerChannel parameter specifies the YouTube channel ID of the channel to which a video is being added. This parameter is required when a request specifies a value for the onBehalfOfContentOwner parameter, and it can only be used in conjunction with that parameter. In addition, the request must be authorized using a CMS account that is linked to the content owner that the onBehalfOfContentOwner parameter specifies. Finally, the channel that the onBehalfOfContentOwnerChannel parameter value specifies must be linked to the content owner that the onBehalfOfContentOwner parameter specifies.

This parameter is intended for YouTube content partners that own and manage many different YouTube channels. It allows content owners to authenticate once and perform actions on behalf of the channel specified in the parameter value, without having to provide authentication credentials for each separate channel.

Sets the on behalf of content owner channel query property to the given value.

fn on_behalf_of_content_owner(self, new_value: &str) -> LiveStreamInsertCall<'a, C, A>

Note: This parameter is intended exclusively for YouTube content partners.

The onBehalfOfContentOwner parameter indicates that the request's authorization credentials identify a YouTube CMS user who is acting on behalf of the content owner specified in the parameter value. This parameter is intended for YouTube content partners that own and manage many different YouTube channels. It allows content owners to authenticate once and get access to all their video and channel data, without having to provide authentication credentials for each individual channel. The CMS account that the user authenticates with must be linked to the specified YouTube content owner.

Sets the on behalf of content owner query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> LiveStreamInsertCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> LiveStreamInsertCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> LiveStreamInsertCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.