Struct google_youtube3::MonitorStreamInfo

pub struct MonitorStreamInfo {
    pub broadcast_stream_delay_ms: Option<u32>,
    pub embed_html: Option<String>,
    pub enable_monitor_stream: Option<bool>,
}

Settings and Info of the monitor stream

This type is not used in any activity, and only used as part of another schema.

Fields
broadcast_stream_delay_ms: Option<u32>

If you have set the enableMonitorStream property to true, then this property determines the length of the live broadcast delay.

embed_html: Option<String>

HTML code that embeds a player that plays the monitor stream.

enable_monitor_stream: Option<bool>

This value determines whether the monitor stream is enabled for the broadcast. If the monitor stream is enabled, then YouTube will broadcast the event content on a special stream intended only for the broadcaster's consumption. The broadcaster can use the stream to review the event content and also to identify the optimal times to insert cuepoints.

You need to set this value to true if you intend to have a broadcast delay for your event.

Note: This property cannot be updated once the broadcast is in the testing or live state.