Struct google_youtube3::PlaylistItem

pub struct PlaylistItem {
    pub status: Option<PlaylistItemStatus>,
    pub snippet: Option<PlaylistItemSnippet>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub content_details: Option<PlaylistItemContentDetails>,
    pub id: Option<String>,
}

A playlistItem resource identifies another resource, such as a video, that is included in a playlist. In addition, the playlistItem resource contains details about the included resource that pertain specifically to how that resource is used in that playlist.

YouTube uses playlists to identify special collections of videos for a channel, such as:
- uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information.

You can retrieve the playlist IDs for each of these lists from the channel resource for a given channel. You can then use the playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the playlistItems.insert and playlistItems.delete methods. For example, if a user gives a positive rating to a video, you would insert that video into the liked videos playlist for that user's channel.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert playlist items (request|response)

update playlist items (request|response)

list playlist items (none)

delete playlist items (none)

Fields
status: Option<PlaylistItemStatus>

The status object contains information about the playlist item's privacy status.

snippet: Option<PlaylistItemSnippet>

The snippet object contains basic details about the playlist item, such as its title and position in the playlist.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "youtube#playlistItem".

etag: Option<String>

Etag of this resource.

content_details: Option<PlaylistItemContentDetails>

The contentDetails object is included in the resource if the included item is a YouTube video. The object contains additional information about the video.

id: Option<String>

The ID that YouTube uses to uniquely identify the playlist item.