Struct google_youtube3::PlaylistItemMethods

pub struct PlaylistItemMethods<'a, C, A> where C: 'a, A: 'a { /* fields omitted */ }

A builder providing access to all methods supported on playlistItem resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `insert(...)`, `list(...)` and `update(...)`
// to build up your call.
let rb = hub.playlist_items();

Methods

impl<'a, C, A> PlaylistItemMethods<'a, C, A>

fn insert(&self, request: PlaylistItem) -> PlaylistItemInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a resource to a playlist.

Arguments

request - No description provided.

fn delete(&self, id: &str) -> PlaylistItemDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a playlist item.

Arguments

id - The id parameter specifies the YouTube playlist item ID for the playlist item that is being deleted. In a playlistItem resource, the id property specifies the playlist item's ID.

fn list(&self, part: &str) -> PlaylistItemListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a collection of playlist items that match the API request parameters. You can retrieve all of the playlist items in a specified playlist or retrieve one or more playlist items by their unique IDs.

Arguments

part - The part parameter specifies a comma-separated list of one or more playlistItem resource properties that the API response will include. If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in a playlistItem resource, the snippet property contains numerous fields, including the title, description, position, and resourceId properties. As such, if you set part=snippet, the API response will contain all of those properties.

fn update(&self, request: PlaylistItem) -> PlaylistItemUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies a playlist item. For example, you could update the item's position in the playlist.

Arguments

request - No description provided.