Struct google_youtube3::PlaylistMethods

pub struct PlaylistMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on playlist resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `insert(...)`, `list(...)` and `update(...)`
// to build up your call.
let rb = hub.playlists();

Methods

impl<'a, C, A> PlaylistMethods<'a, C, A>

fn insert(&self, request: Playlist) -> PlaylistInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a playlist.

Arguments

request - No description provided.

fn list(&self, part: &str) -> PlaylistListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a collection of playlists that match the API request parameters. For example, you can retrieve all playlists that the authenticated user owns, or you can retrieve one or more playlists by their unique IDs.

Arguments

part - The part parameter specifies a comma-separated list of one or more playlist resource properties that the API response will include. The part names that you can include in the parameter value are id, snippet, status, and contentDetails. If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in a playlist resource, the snippet property contains properties like author, title, description, tags, and timeCreated. As such, if you set part=snippet, the API response will contain all of those properties.

fn delete(&self, id: &str) -> PlaylistDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a playlist.

Arguments

id - The id parameter specifies the YouTube playlist ID for the playlist that is being deleted. In a playlist resource, the id property specifies the playlist's ID.

fn update(&self, request: Playlist) -> PlaylistUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies a playlist. For example, you could change a playlist's title, description, or privacy status.

Arguments

request - No description provided.