Struct google_youtube3::PlaylistSnippet

pub struct PlaylistSnippet {
    pub description: Option<String>,
    pub title: Option<String>,
    pub channel_id: Option<String>,
    pub published_at: Option<String>,
    pub tags: Option<Vec<String>>,
    pub channel_title: Option<String>,
    pub default_language: Option<String>,
    pub localized: Option<PlaylistLocalization>,
    pub thumbnails: Option<ThumbnailDetails>,
}

Basic details about a playlist, including title, description and thumbnails.

This type is not used in any activity, and only used as part of another schema.

Fields

description

The playlist's description.

title

The playlist's title.

channel_id

The ID that YouTube uses to uniquely identify the channel that published the playlist.

published_at

The date and time that the playlist was created. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

tags

Keyword tags associated with the playlist.

channel_title

The channel title of the channel that the video belongs to.

default_language

The language of the playlist's default title and description.

localized

Localized title and description, read-only.

thumbnails

A map of thumbnail images associated with the playlist. For each object in the map, the key is the name of the thumbnail image, and the value is an object that contains other information about the thumbnail.