Struct google_youtube3::SearchResultSnippet

pub struct SearchResultSnippet {
    pub live_broadcast_content: Option<String>,
    pub description: Option<String>,
    pub title: Option<String>,
    pub thumbnails: Option<ThumbnailDetails>,
    pub channel_id: Option<String>,
    pub published_at: Option<String>,
    pub channel_title: Option<String>,
}

Basic details about a search result, including title, description and thumbnails of the item referenced by the search result.

This type is not used in any activity, and only used as part of another schema.

Fields

live_broadcast_content

It indicates if the resource (video or channel) has upcoming/active live broadcast content. Or it's "none" if there is not any upcoming/active live broadcasts.

description

A description of the search result.

title

The title of the search result.

thumbnails

A map of thumbnail images associated with the search result. For each object in the map, the key is the name of the thumbnail image, and the value is an object that contains other information about the thumbnail.

channel_id

The value that YouTube uses to uniquely identify the channel that published the resource that the search result identifies.

published_at

The creation date and time of the resource that the search result identifies. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

channel_title

The title of the channel that published the resource that the search result identifies.