Struct google_youtube3::SubscriptionSnippet

pub struct SubscriptionSnippet {
    pub description: Option<String>,
    pub title: Option<String>,
    pub resource_id: Option<ResourceId>,
    pub thumbnails: Option<ThumbnailDetails>,
    pub channel_id: Option<String>,
    pub published_at: Option<String>,
    pub channel_title: Option<String>,
}

Basic details about a subscription, including title, description and thumbnails of the subscribed item.

This type is not used in any activity, and only used as part of another schema.

Fields

description

The subscription's details.

title

The subscription's title.

resource_id

The id object contains information about the channel that the user subscribed to.

thumbnails

A map of thumbnail images associated with the video. For each object in the map, the key is the name of the thumbnail image, and the value is an object that contains other information about the thumbnail.

channel_id

The ID that YouTube uses to uniquely identify the subscriber's channel.

published_at

The date and time that the subscription was created. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

channel_title

Channel title for the channel that the subscription belongs to.