Struct google_youtube3::VideoFileDetails

pub struct VideoFileDetails {
    pub bitrate_bps: Option<String>,
    pub container: Option<String>,
    pub recording_location: Option<GeoPoint>,
    pub file_type: Option<String>,
    pub creation_time: Option<String>,
    pub duration_ms: Option<String>,
    pub file_name: Option<String>,
    pub file_size: Option<String>,
    pub video_streams: Option<Vec<VideoFileDetailsVideoStream>>,
    pub audio_streams: Option<Vec<VideoFileDetailsAudioStream>>,
}

Describes original video file properties, including technical details about audio and video streams, but also metadata information like content length, digitization time, or geotagging information.

This type is not used in any activity, and only used as part of another schema.

Fields

bitrate_bps

The uploaded video file's combined (video and audio) bitrate in bits per second.

container

The uploaded video file's container format.

recording_location

Geographic coordinates that identify the place where the uploaded video was recorded. Coordinates are defined using WGS 84.

file_type

The uploaded file's type as detected by YouTube's video processing engine. Currently, YouTube only processes video files, but this field is present whether a video file or another type of file was uploaded.

creation_time

The date and time when the uploaded video file was created. The value is specified in ISO 8601 format. Currently, the following ISO 8601 formats are supported: - Date only: YYYY-MM-DD - Naive time: YYYY-MM-DDTHH:MM:SS - Time with timezone: YYYY-MM-DDTHH:MM:SS+HH:MM

duration_ms

The length of the uploaded video in milliseconds.

file_name

The uploaded file's name. This field is present whether a video file or another type of file was uploaded.

file_size

The uploaded file's size in bytes. This field is present whether a video file or another type of file was uploaded.

video_streams

A list of video streams contained in the uploaded video file. Each item in the list contains detailed metadata about a video stream.

audio_streams

A list of audio streams contained in the uploaded video file. Each item in the list contains detailed metadata about an audio stream.