Struct google_youtube3::VideoFileDetailsVideoStream

pub struct VideoFileDetailsVideoStream {
    pub bitrate_bps: Option<String>,
    pub vendor: Option<String>,
    pub codec: Option<String>,
    pub width_pixels: Option<u32>,
    pub height_pixels: Option<u32>,
    pub aspect_ratio: Option<f64>,
    pub rotation: Option<String>,
    pub frame_rate_fps: Option<f64>,
}

Information about a video stream.

This type is not used in any activity, and only used as part of another schema.

Fields

bitrate_bps

The video stream's bitrate, in bits per second.

vendor

A value that uniquely identifies a video vendor. Typically, the value is a four-letter vendor code.

codec

The video codec that the stream uses.

width_pixels

The encoded video content's width in pixels. You can calculate the video's encoding aspect ratio as width_pixels / height_pixels.

height_pixels

The encoded video content's height in pixels.

aspect_ratio

The video content's display aspect ratio, which specifies the aspect ratio in which the video should be displayed.

rotation

The amount that YouTube needs to rotate the original source content to properly display the video.

frame_rate_fps

The video stream's frame rate, in frames per second.