Struct google_youtube3::VideoInsertCall

pub struct VideoInsertCall<'a, C, A> where C: 'a, A: 'a { /* fields omitted */ }

Uploads a video to YouTube and optionally sets the video's metadata.

A builder for the insert method supported by a video resource. It is not used directly, but through a VideoMethods instance.

Example

Instantiate a resource method builder

use youtube3::Video;
use std::fs;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Video::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload_resumable(...)`.
// Values shown here are possibly random and not representative !
let result = hub.videos().insert(req)
             .stabilize(true)
             .on_behalf_of_content_owner_channel("clita")
             .on_behalf_of_content_owner("consetetur")
             .notify_subscribers(false)
             .auto_levels(true)
             .upload_resumable(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());

Methods

impl<'a, C, A> VideoInsertCall<'a, C, A> where C: BorrowMut<Client>,
A: GetToken

fn upload<RS>(self, stream: RS, mime_type: Mime) -> Result<(Response, Video)> where RS: ReadSeek

Upload media all at once. If the upload fails for whichever reason, all progress is lost.

max size: 64GB

multipart: yes

valid mime types: 'application/octet-stream' and 'video/*'

fn upload_resumable<RS>(self,
resumeable_stream: RS,
mime_type: Mime)
-> Result<(Response, Video)> where RS: ReadSeek

Upload media in a resumable fashion. Even if the upload fails or is interrupted, it can be resumed for a certain amount of time as the server maintains state temporarily.

The delegate will be asked for an upload_url(), and if not provided, will be asked to store an upload URL that was provided by the server, using store_upload_url(...). The upload will be done in chunks, the delegate may specify the chunk_size() and may cancel the operation before each chunk is uploaded, using cancel_chunk_upload(...).

max size: 64GB

multipart: yes

valid mime types: 'application/octet-stream' and 'video/*'

fn request(self, new_value: Video) -> VideoInsertCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn part(self, new_value: &str) -> VideoInsertCall<'a, C, A>

The part parameter serves two purposes in this operation. It identifies the properties that the write operation will set as well as the properties that the API response will include.

Note that not all parts contain properties that can be set when inserting or updating a video. For example, the statistics object encapsulates statistics that YouTube calculates for a video and does not contain values that you can set or modify. If the parameter value specifies a part that does not contain mutable values, that part will still be included in the API response.

Sets the part query property to the given value.

Even though the parts list is automatically derived from Resource passed in during instantiation and indicates which values you are passing, the response would contain the very same parts. This may not always be desirable, as you can obtain (newly generated) parts you cannot pass in, like statistics that are generated server side. Therefore you should use this method to specify the parts you provide in addition to the ones you want in the response.

fn stabilize(self, new_value: bool) -> VideoInsertCall<'a, C, A>

The stabilize parameter indicates whether YouTube should adjust the video to remove shaky camera motions.

Sets the stabilize query property to the given value.

fn on_behalf_of_content_owner_channel(self,
new_value: &str)
-> VideoInsertCall<'a, C, A>

This parameter can only be used in a properly authorized request. Note: This parameter is intended exclusively for YouTube content partners.

The onBehalfOfContentOwnerChannel parameter specifies the YouTube channel ID of the channel to which a video is being added. This parameter is required when a request specifies a value for the onBehalfOfContentOwner parameter, and it can only be used in conjunction with that parameter. In addition, the request must be authorized using a CMS account that is linked to the content owner that the onBehalfOfContentOwner parameter specifies. Finally, the channel that the onBehalfOfContentOwnerChannel parameter value specifies must be linked to the content owner that the onBehalfOfContentOwner parameter specifies.

This parameter is intended for YouTube content partners that own and manage many different YouTube channels. It allows content owners to authenticate once and perform actions on behalf of the channel specified in the parameter value, without having to provide authentication credentials for each separate channel.

Sets the on behalf of content owner channel query property to the given value.

fn on_behalf_of_content_owner(self,
new_value: &str)
-> VideoInsertCall<'a, C, A>

Note: This parameter is intended exclusively for YouTube content partners.

The onBehalfOfContentOwner parameter indicates that the request's authorization credentials identify a YouTube CMS user who is acting on behalf of the content owner specified in the parameter value. This parameter is intended for YouTube content partners that own and manage many different YouTube channels. It allows content owners to authenticate once and get access to all their video and channel data, without having to provide authentication credentials for each individual channel. The CMS account that the user authenticates with must be linked to the specified YouTube content owner.

Sets the on behalf of content owner query property to the given value.

fn notify_subscribers(self, new_value: bool) -> VideoInsertCall<'a, C, A>

The notifySubscribers parameter indicates whether YouTube should send a notification about the new video to users who subscribe to the video's channel. A parameter value of True indicates that subscribers will be notified of newly uploaded videos. However, a channel owner who is uploading many videos might prefer to set the value to False to avoid sending a notification about each new video to the channel's subscribers.

Sets the notify subscribers query property to the given value.

fn auto_levels(self, new_value: bool) -> VideoInsertCall<'a, C, A>

The autoLevels parameter indicates whether YouTube should automatically enhance the video's lighting and color.

Sets the auto levels query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> VideoInsertCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> VideoInsertCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> VideoInsertCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.