Struct google_youtube3::VideoMethods

pub struct VideoMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on video resources. It is not used directly, but through the YouTube hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get_rating(...)`, `insert(...)`, `list(...)`, `rate(...)`, `report_abuse(...)` and `update(...)`
// to build up your call.
let rb = hub.videos();

Methods

impl<'a, C, A> VideoMethods<'a, C, A>

fn report_abuse(&self, request: VideoAbuseReport) -> VideoReportAbuseCall<'a, C, A>

Create a builder to help you perform the following task:

Report abuse for a video.

Arguments

request - No description provided.

fn list(&self, part: &str) -> VideoListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of videos that match the API request parameters.

Arguments

part - The part parameter specifies a comma-separated list of one or more video resource properties that the API response will include. If the parameter identifies a property that contains child properties, the child properties will be included in the response. For example, in a video resource, the snippet property contains the channelId, title, description, tags, and categoryId properties. As such, if you set part=snippet, the API response will contain all of those properties.

fn rate(&self, id: &str, rating: &str) -> VideoRateCall<'a, C, A>

Create a builder to help you perform the following task:

Add a like or dislike rating to a video or remove a rating from a video.

Arguments

id - The id parameter specifies the YouTube video ID of the video that is being rated or having its rating removed.

rating - Specifies the rating to record.

fn get_rating(&self, id: &str) -> VideoGetRatingCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the ratings that the authorized user gave to a list of specified videos.

Arguments

id - The id parameter specifies a comma-separated list of the YouTube video ID(s) for the resource(s) for which you are retrieving rating data. In a video resource, the id property specifies the video's ID.

fn delete(&self, id: &str) -> VideoDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a YouTube video.

Arguments

id - The id parameter specifies the YouTube video ID for the resource that is being deleted. In a video resource, the id property specifies the video's ID.

fn update(&self, request: Video) -> VideoUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a video's metadata.

Arguments

request - No description provided.

fn insert(&self, request: Video) -> VideoInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads a video to YouTube and optionally sets the video's metadata.

Arguments

request - No description provided.