Struct google_youtube3::VideoProcessingDetails

pub struct VideoProcessingDetails {
    pub file_details_availability: Option<String>,
    pub editor_suggestions_availability: Option<String>,
    pub processing_status: Option<String>,
    pub processing_issues_availability: Option<String>,
    pub processing_failure_reason: Option<String>,
    pub thumbnails_availability: Option<String>,
    pub processing_progress: Option<VideoProcessingDetailsProcessingProgress>,
    pub tag_suggestions_availability: Option<String>,
}

Describes processing status and progress and availability of some other Video resource parts.

This type is not used in any activity, and only used as part of another schema.

Fields

file_details_availability

This value indicates whether file details are available for the uploaded video. You can retrieve a video's file details by requesting the fileDetails part in your videos.list() request.

editor_suggestions_availability

This value indicates whether video editing suggestions, which might improve video quality or the playback experience, are available for the video. You can retrieve these suggestions by requesting the suggestions part in your videos.list() request.

processing_status

The video's processing status. This value indicates whether YouTube was able to process the video or if the video is still being processed.

processing_issues_availability

This value indicates whether the video processing engine has generated suggestions that might improve YouTube's ability to process the the video, warnings that explain video processing problems, or errors that cause video processing problems. You can retrieve these suggestions by requesting the suggestions part in your videos.list() request.

processing_failure_reason

The reason that YouTube failed to process the video. This property will only have a value if the processingStatus property's value is failed.

thumbnails_availability

This value indicates whether thumbnail images have been generated for the video.

processing_progress

The processingProgress object contains information about the progress YouTube has made in processing the video. The values are really only relevant if the video's processing status is processing.

tag_suggestions_availability

This value indicates whether keyword (tag) suggestions are available for the video. Tags can be added to a video's metadata to make it easier for other users to find the video. You can retrieve these suggestions by requesting the suggestions part in your videos.list() request.