Struct google_youtube3::VideoStatus

pub struct VideoStatus {
    pub license: Option<String>,
    pub embeddable: Option<bool>,
    pub privacy_status: Option<String>,
    pub publish_at: Option<String>,
    pub public_stats_viewable: Option<bool>,
    pub upload_status: Option<String>,
    pub rejection_reason: Option<String>,
    pub failure_reason: Option<String>,
}

Basic details about a video category, such as its localized title.

This type is not used in any activity, and only used as part of another schema.

Fields
license: Option<String>

The video's license.

embeddable: Option<bool>

This value indicates if the video can be embedded on another website.

privacy_status: Option<String>

The video's privacy status.

publish_at: Option<String>

The date and time when the video is scheduled to publish. It can be set only if the privacy status of the video is private. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

public_stats_viewable: Option<bool>

This value indicates if the extended video statistics on the watch page can be viewed by everyone. Note that the view count, likes, etc will still be visible if this is disabled.

upload_status: Option<String>

The status of the uploaded video.

rejection_reason: Option<String>

This value explains why YouTube rejected an uploaded video. This property is only present if the uploadStatus property indicates that the upload was rejected.

failure_reason: Option<String>

This value explains why a video failed to upload. This property is only present if the uploadStatus property indicates that the upload failed.