Crate google_youtube3

This documentation was generated from YouTube crate version 5.0.3+20240303, where 20240303 is the exact revision of the youtube:v3 schema built by the mako code generator v5.0.3.

Everything else about the YouTube v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub …

abuse reports

insert

activities

list

captions

delete, download, insert, list and update

channel banners

insert

channel sections

delete, insert, list and update

channels

list and update

comment threads

insert and list

comments

delete, insert, list, mark as spam, set moderation status and update

i18n languages

list

i18n regions

list

live broadcasts

bind, delete, insert, insert cuepoint, list, transition and update

live chat bans

delete and insert

live chat messages

delete, insert and list

live chat moderators

delete, insert and list

live streams

delete, insert, list and update

members

list

memberships levels

list

playlist images

delete, insert, list and update

playlist items

delete, insert, list and update

playlists

delete, insert, list and update

search

list

subscriptions

delete, insert and list

super chat events

list

tests

insert

third party links

delete, insert, list and update

thumbnails

set

video abuse report reasons

list

video categories

list

videos

delete, get rating, insert, list, rate, report abuse and update

watermarks

set and unset

youtube

v3 update comment threads

Upload supported by …

insert captions

update captions

insert channel banners

insert playlist images

update playlist images

set thumbnails

insert videos

set watermarks

Download supported by …

download captions

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically …

let r = hub.videos().delete(...).doit().await
let r = hub.videos().get_rating(...).doit().await
let r = hub.videos().insert(...).doit().await
let r = hub.videos().list(...).doit().await
let r = hub.videos().rate(...).doit().await
let r = hub.videos().report_abuse(...).doit().await
let r = hub.videos().update(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-youtube3 = "*"
serde = "^1.0"
serde_json = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate google_youtube3 as youtube3;
use youtube3::{Result, Error};
use std::default::Default;
use youtube3::{YouTube, oauth2, hyper, hyper_rustls, chrono, FieldMask};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = oauth2::InstalledFlowAuthenticator::builder(
        secret,
        oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = YouTube::new(hyper::Client::builder().build(hyper_rustls::HttpsConnectorBuilder::new().with_native_roots().https_or_http().enable_http1().build()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.videos().list(&vec!["gubergren".into()])
             .video_category_id("Lorem")
             .region_code("gubergren")
             .page_token("eos")
             .on_behalf_of_content_owner("dolor")
             .my_rating("ea")
             .max_width(-55)
             .max_results(13)
             .max_height(-47)
             .locale("duo")
             .add_id("ipsum")
             .hl("sed")
             .chart("ut")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it’s methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub extern crate google_apis_common as client;

pub use api::YouTube;

pub use hyper;

pub use hyper_rustls;

pub use client::chrono;

pub use client::oauth2;

Modules

api

Structs

FieldMask

A FieldMask as defined in https://github.com/protocolbuffers/protobuf/blob/ec1a70913e5793a7d0a7b5fbf7e0e4f75409dd41/src/google/protobuf/field_mask.proto#L180

Enums

Error

Traits

Delegate

A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Type Aliases

Result

A universal result type used as return for all calls.