Module google_youtube3::api

Structs

AbuseReport
There is no detailed description.

AbuseReportInsertCall
Inserts a new resource into this collection.

AbuseReportMethods
A builder providing access to all methods supported on abuseReport resources. It is not used directly, but through the YouTube hub.

AbuseType
There is no detailed description.

AccessPolicy
Rights management policy for YouTube resources.

Activity
An activity resource contains information about an action that a particular channel, or user, has taken on YouTube.The actions reported in activity feeds include rating a video, sharing a video, marking a video as a favorite, commenting on a video, uploading a video, and so forth. Each activity resource identifies the type of action, the channel associated with the action, and the resource(s) associated with the action, such as the video that was rated or uploaded.

ActivityContentDetails
Details about the content of an activity: the video that was shared, the channel that was subscribed to, etc.

ActivityContentDetailsBulletin
Details about a channel bulletin post.

ActivityContentDetailsChannelItem
Details about a resource which was added to a channel.

ActivityContentDetailsComment
Information about a resource that received a comment.

ActivityContentDetailsFavorite
Information about a video that was marked as a favorite video.

ActivityContentDetailsLike
Information about a resource that received a positive (like) rating.

ActivityContentDetailsPlaylistItem
Information about a new playlist item.

ActivityContentDetailsPromotedItem
Details about a resource which is being promoted.

ActivityContentDetailsRecommendation
Information that identifies the recommended resource.

ActivityContentDetailsSocial
Details about a social network post.

ActivityContentDetailsSubscription
Information about a channel that a user subscribed to.

ActivityContentDetailsUpload
Information about the uploaded video.

ActivityListCall
Retrieves a list of resources, possibly filtered.

ActivityListResponse
There is no detailed description.

ActivityMethods
A builder providing access to all methods supported on activity resources. It is not used directly, but through the YouTube hub.

ActivitySnippet
Basic details about an activity, including title, description, thumbnails, activity type and group. Next ID: 12

Caption
A caption resource represents a YouTube caption track. A caption track is associated with exactly one YouTube video.

CaptionDeleteCall
Deletes a resource.

CaptionDownloadCall
Downloads a caption track.

CaptionInsertCall
Inserts a new resource into this collection.

CaptionListCall
Retrieves a list of resources, possibly filtered.

CaptionListResponse
There is no detailed description.

CaptionMethods
A builder providing access to all methods supported on caption resources. It is not used directly, but through the YouTube hub.

CaptionSnippet
Basic details about a caption track, such as its language and name.

CaptionUpdateCall
Updates an existing resource.

CdnSettings
Brief description of the live stream cdn settings.

Channel
A channel resource contains information about a YouTube channel.

ChannelAuditDetails
The auditDetails object encapsulates channel data that is relevant for YouTube Partners during the audit process.

ChannelBannerInsertCall
Inserts a new resource into this collection.

ChannelBannerMethods
A builder providing access to all methods supported on channelBanner resources. It is not used directly, but through the YouTube hub.

ChannelBannerResource
A channel banner returned as the response to a channel_banner.insert call.

ChannelBrandingSettings
Branding properties of a YouTube channel.

ChannelContentDetails
Details about the content of a channel.

ChannelContentDetailsRelatedPlaylists
There is no detailed description.

ChannelContentOwnerDetails
The contentOwnerDetails object encapsulates channel data that is relevant for YouTube Partners linked with the channel.

ChannelConversionPing
Pings that the app shall fire (authenticated by biscotti cookie). Each ping has a context, in which the app must fire the ping, and a url identifying the ping.

ChannelConversionPings
The conversionPings object encapsulates information about conversion pings that need to be respected by the channel.

ChannelListCall
Retrieves a list of resources, possibly filtered.

ChannelListResponse
There is no detailed description.

ChannelLocalization
Channel localization setting

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the YouTube hub.

ChannelProfileDetails
There is no detailed description.

ChannelSection
There is no detailed description.

ChannelSectionContentDetails
Details about a channelsection, including playlists and channels.

ChannelSectionDeleteCall
Deletes a resource.

ChannelSectionInsertCall
Inserts a new resource into this collection.

ChannelSectionListCall
Retrieves a list of resources, possibly filtered.

ChannelSectionListResponse
There is no detailed description.

ChannelSectionLocalization
ChannelSection localization setting

ChannelSectionMethods
A builder providing access to all methods supported on channelSection resources. It is not used directly, but through the YouTube hub.

ChannelSectionSnippet
Basic details about a channel section, including title, style and position.

ChannelSectionTargeting
ChannelSection targeting setting.

ChannelSectionUpdateCall
Updates an existing resource.

ChannelSettings
Branding properties for the channel view.

ChannelSnippet
Basic details about a channel, including title, description and thumbnails.

ChannelStatistics
Statistics about a channel: number of subscribers, number of videos in the channel, etc.

ChannelStatus
JSON template for the status part of a channel.

ChannelToStoreLinkDetails
Information specific to a store on a merchandising platform linked to a YouTube channel.

ChannelTopicDetails
Freebase topic information related to the channel.

ChannelUpdateCall
Updates an existing resource.

Comment
A comment represents a single YouTube comment.

CommentDeleteCall
Deletes a resource.

CommentInsertCall
Inserts a new resource into this collection.

CommentListCall
Retrieves a list of resources, possibly filtered.

CommentListResponse
There is no detailed description.

CommentMarkAsSpamCall
Expresses the caller's opinion that one or more comments should be flagged as spam.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the YouTube hub.

CommentSetModerationStatuCall
Sets the moderation status of one or more comments.

CommentSnippet
Basic details about a comment, such as its author and text.

CommentSnippetAuthorChannelId
The id of the author's YouTube channel, if any.

CommentThread
A comment thread represents information that applies to a top level comment and all its replies. It can also include the top level comment itself and some of the replies.

CommentThreadInsertCall
Inserts a new resource into this collection.

CommentThreadListCall
Retrieves a list of resources, possibly filtered.

CommentThreadListResponse
There is no detailed description.

CommentThreadMethods
A builder providing access to all methods supported on commentThread resources. It is not used directly, but through the YouTube hub.

CommentThreadReplies
Comments written in (direct or indirect) reply to the top level comment.

CommentThreadSnippet
Basic details about a comment thread.

CommentThreadUpdateCall
Updates an existing resource.

CommentUpdateCall
Updates an existing resource.

ContentRating
Ratings schemes. The country-specific ratings are mostly for movies and shows. LINT.IfChange

Entity
There is no detailed description.

GeoPoint
Geographical coordinates of a point, in WGS84.

I18nLanguage
An i18nLanguage resource identifies a UI language currently supported by YouTube.

I18nLanguageListCall
Retrieves a list of resources, possibly filtered.

I18nLanguageListResponse
There is no detailed description.

I18nLanguageMethods
A builder providing access to all methods supported on i18nLanguage resources. It is not used directly, but through the YouTube hub.

I18nLanguageSnippet
Basic details about an i18n language, such as language code and human-readable name.

I18nRegion
A i18nRegion resource identifies a region where YouTube is available.

I18nRegionListCall
Retrieves a list of resources, possibly filtered.

I18nRegionListResponse
There is no detailed description.

I18nRegionMethods
A builder providing access to all methods supported on i18nRegion resources. It is not used directly, but through the YouTube hub.

I18nRegionSnippet
Basic details about an i18n region, such as region code and human-readable name.

ImageSettings
Branding properties for images associated with the channel.

IngestionInfo
Describes information necessary for ingesting an RTMP or an HTTP stream.

InvideoBranding
LINT.IfChange Describes an invideo branding.

InvideoPosition
Describes the spatial position of a visual widget inside a video. It is a union of various position types, out of which only will be set one.

InvideoTiming
Describes a temporal position of a visual widget inside a video.

LanguageTag
There is no detailed description.

LevelDetails
There is no detailed description.

LiveBroadcast
A liveBroadcast resource represents an event that will be streamed, via live video, on YouTube.

LiveBroadcastBindCall
Bind a broadcast to a stream.

LiveBroadcastContentDetails
Detailed settings of a broadcast.

LiveBroadcastDeleteCall
Delete a given broadcast.

LiveBroadcastInsertCall
Inserts a new stream for the authenticated user.

LiveBroadcastListCall
Retrieve the list of broadcasts associated with the given channel.

LiveBroadcastListResponse
There is no detailed description.

LiveBroadcastMethods
A builder providing access to all methods supported on liveBroadcast resources. It is not used directly, but through the YouTube hub.

LiveBroadcastSnippet
Basic broadcast information.

LiveBroadcastStatistics
Statistics about the live broadcast. These represent a snapshot of the values at the time of the request. Statistics are only returned for live broadcasts.

LiveBroadcastStatus
Live broadcast state.

LiveBroadcastTransitionCall
Transition a broadcast to a given status.

LiveBroadcastUpdateCall
Updates an existing broadcast for the authenticated user.

LiveChatBan
A __liveChatBan__ resource represents a ban for a YouTube live chat.

LiveChatBanDeleteCall
Deletes a chat ban.

LiveChatBanInsertCall
Inserts a new resource into this collection.

LiveChatBanMethods
A builder providing access to all methods supported on liveChatBan resources. It is not used directly, but through the YouTube hub.

LiveChatBanSnippet
There is no detailed description.

LiveChatFanFundingEventDetails
There is no detailed description.

LiveChatMessage
A liveChatMessage resource represents a chat message in a YouTube Live Chat.

LiveChatMessageAuthorDetails
There is no detailed description.

LiveChatMessageDeleteCall
Deletes a chat message.

LiveChatMessageDeletedDetails
There is no detailed description.

LiveChatMessageInsertCall
Inserts a new resource into this collection.

LiveChatMessageListCall
Retrieves a list of resources, possibly filtered.

LiveChatMessageListResponse
There is no detailed description.

LiveChatMessageMethods
A builder providing access to all methods supported on liveChatMessage resources. It is not used directly, but through the YouTube hub.

LiveChatMessageRetractedDetails
There is no detailed description.

LiveChatMessageSnippet
There is no detailed description.

LiveChatModerator
A liveChatModerator resource represents a moderator for a YouTube live chat. A chat moderator has the ability to ban/unban users from a chat, remove message, etc.

LiveChatModeratorDeleteCall
Deletes a chat moderator.

LiveChatModeratorInsertCall
Inserts a new resource into this collection.

LiveChatModeratorListCall
Retrieves a list of resources, possibly filtered.

LiveChatModeratorListResponse
There is no detailed description.

LiveChatModeratorMethods
A builder providing access to all methods supported on liveChatModerator resources. It is not used directly, but through the YouTube hub.

LiveChatModeratorSnippet
There is no detailed description.

LiveChatSuperChatDetails
There is no detailed description.

LiveChatSuperStickerDetails
There is no detailed description.

LiveChatTextMessageDetails
There is no detailed description.

LiveChatUserBannedMessageDetails
There is no detailed description.

LiveStream
A live stream describes a live ingestion point.

LiveStreamConfigurationIssue
There is no detailed description.

LiveStreamContentDetails
Detailed settings of a stream.

LiveStreamDeleteCall
Deletes an existing stream for the authenticated user.

LiveStreamHealthStatus
There is no detailed description.

LiveStreamInsertCall
Inserts a new stream for the authenticated user.

LiveStreamListCall
Retrieve the list of streams associated with the given channel. --

LiveStreamListResponse
There is no detailed description.

LiveStreamMethods
A builder providing access to all methods supported on liveStream resources. It is not used directly, but through the YouTube hub.

LiveStreamSnippet
There is no detailed description.

LiveStreamStatus
Brief description of the live stream status.

LiveStreamUpdateCall
Updates an existing stream for the authenticated user.

LocalizedProperty
There is no detailed description.

LocalizedString
There is no detailed description.

Member
A member resource represents a member for a YouTube channel. A member provides recurring monetary support to a creator and receives special benefits.

MemberListCall
Retrieves a list of members that match the request criteria for a channel.

MemberListResponse
There is no detailed description.

MemberMethods
A builder providing access to all methods supported on member resources. It is not used directly, but through the YouTube hub.

MemberSnippet
There is no detailed description.

MembershipsDetails
There is no detailed description.

MembershipsDuration
There is no detailed description.

MembershipsDurationAtLevel
There is no detailed description.

MembershipsLevel
A membershipsLevel resource represents an offer made by YouTube creators for their fans. Users can become members of the channel by joining one of the available levels. They will provide recurring monetary support and receives special benefits.

MembershipsLevelListCall
Retrieves a list of all pricing levels offered by a creator to the fans.

MembershipsLevelListResponse
There is no detailed description.

MembershipsLevelMethods
A builder providing access to all methods supported on membershipsLevel resources. It is not used directly, but through the YouTube hub.

MembershipsLevelSnippet
There is no detailed description.

MonitorStreamInfo
Settings and Info of the monitor stream

PageInfo
Paging details for lists of resources, including total number of items available and number of resources returned in a single page.

Playlist
A playlist resource represents a YouTube playlist. A playlist is a collection of videos that can be viewed sequentially and shared with other users. A playlist can contain up to 200 videos, and YouTube does not limit the number of playlists that each user creates. By default, playlists are publicly visible to other users, but playlists can be public or private. YouTube also uses playlists to identify special collections of videos for a channel, such as: - uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information. You can retrieve the playlist IDs for each of these lists from the channel resource for a given channel. You can then use the playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the playlistItems.insert and playlistItems.delete methods.

PlaylistContentDetails
There is no detailed description.

PlaylistDeleteCall
Deletes a resource.

PlaylistInsertCall
Inserts a new resource into this collection.

PlaylistItem
A playlistItem resource identifies another resource, such as a video, that is included in a playlist. In addition, the playlistItem resource contains details about the included resource that pertain specifically to how that resource is used in that playlist. YouTube uses playlists to identify special collections of videos for a channel, such as: - uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information. You can retrieve the playlist IDs for each of these lists from the channel resource for a given channel. You can then use the playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the playlistItems.insert and playlistItems.delete methods. For example, if a user gives a positive rating to a video, you would insert that video into the liked videos playlist for that user's channel.

PlaylistItemContentDetails
There is no detailed description.

PlaylistItemDeleteCall
Deletes a resource.

PlaylistItemInsertCall
Inserts a new resource into this collection.

PlaylistItemListCall
Retrieves a list of resources, possibly filtered.

PlaylistItemListResponse
There is no detailed description.

PlaylistItemMethods
A builder providing access to all methods supported on playlistItem resources. It is not used directly, but through the YouTube hub.

PlaylistItemSnippet
Basic details about a playlist, including title, description and thumbnails. Basic details of a YouTube Playlist item provided by the author. Next ID: 15

PlaylistItemStatus
Information about the playlist item's privacy status.

PlaylistItemUpdateCall
Updates an existing resource.

PlaylistListCall
Retrieves a list of resources, possibly filtered.

PlaylistListResponse
There is no detailed description.

PlaylistLocalization
Playlist localization setting

PlaylistMethods
A builder providing access to all methods supported on playlist resources. It is not used directly, but through the YouTube hub.

PlaylistPlayer
There is no detailed description.

PlaylistSnippet
Basic details about a playlist, including title, description and thumbnails.

PlaylistStatus
There is no detailed description.

PlaylistUpdateCall
Updates an existing resource.

PropertyValue
A pair Property / Value.

RelatedEntity
There is no detailed description.

ResourceId
A resource id is a generic reference that points to another YouTube resource.

SearchListCall
Retrieves a list of search resources

SearchListResponse
There is no detailed description.

SearchMethods
A builder providing access to all methods supported on search resources. It is not used directly, but through the YouTube hub.

SearchResult
A search result contains information about a YouTube video, channel, or playlist that matches the search parameters specified in an API request. While a search result points to a uniquely identifiable resource, like a video, it does not have its own persistent data.

SearchResultSnippet
Basic details about a search result, including title, description and thumbnails of the item referenced by the search result.

Subscription
A subscription resource contains information about a YouTube user subscription. A subscription notifies a user when new videos are added to a channel or when another user takes one of several actions on YouTube, such as uploading a video, rating a video, or commenting on a video.

SubscriptionContentDetails
Details about the content to witch a subscription refers.

SubscriptionDeleteCall
Deletes a resource.

SubscriptionInsertCall
Inserts a new resource into this collection.

SubscriptionListCall
Retrieves a list of resources, possibly filtered.

SubscriptionListResponse
There is no detailed description.

SubscriptionMethods
A builder providing access to all methods supported on subscription resources. It is not used directly, but through the YouTube hub.

SubscriptionSnippet
Basic details about a subscription, including title, description and thumbnails of the subscribed item.

SubscriptionSubscriberSnippet
Basic details about a subscription's subscriber including title, description, channel ID and thumbnails.

SuperChatEvent
A __superChatEvent__ resource represents a Super Chat purchase on a YouTube channel.

SuperChatEventListCall
Retrieves a list of resources, possibly filtered.

SuperChatEventListResponse
There is no detailed description.

SuperChatEventMethods
A builder providing access to all methods supported on superChatEvent resources. It is not used directly, but through the YouTube hub.

SuperChatEventSnippet
There is no detailed description.

SuperStickerMetadata
There is no detailed description.

TestInsertCall
POST method.

TestItem
There is no detailed description.

TestItemTestItemSnippet
There is no detailed description.

TestMethods
A builder providing access to all methods supported on test resources. It is not used directly, but through the YouTube hub.

ThirdPartyLink
A third party account link resource represents a link between a YouTube account or a channel and an account on a third-party service.

ThirdPartyLinkDeleteCall
Deletes a resource.

ThirdPartyLinkInsertCall
Inserts a new resource into this collection.

ThirdPartyLinkListCall
Retrieves a list of resources, possibly filtered.

ThirdPartyLinkMethods
A builder providing access to all methods supported on thirdPartyLink resources. It is not used directly, but through the YouTube hub.

ThirdPartyLinkSnippet
Basic information about a third party account link, including its type and type-specific information.

ThirdPartyLinkStatus
The third-party link status object contains information about the status of the link.

ThirdPartyLinkUpdateCall
Updates an existing resource.

Thumbnail
A thumbnail is an image representing a YouTube resource.

ThumbnailDetails
Internal representation of thumbnails for a YouTube resource.

ThumbnailMethods
A builder providing access to all methods supported on thumbnail resources. It is not used directly, but through the YouTube hub.

ThumbnailSetCall
As this is not an insert in a strict sense (it supports uploading/setting of a thumbnail for multiple videos, which doesn't result in creation of a single resource), I use a custom verb here.

ThumbnailSetResponse
There is no detailed description.

TokenPagination
Stub token pagination template to suppress results.

Video
A video resource represents a YouTube video.

VideoAbuseReport
There is no detailed description.

VideoAbuseReportReason
A __videoAbuseReportReason__ resource identifies a reason that a video could be reported as abusive. Video abuse report reasons are used with video.ReportAbuse.

VideoAbuseReportReasonListCall
Retrieves a list of resources, possibly filtered.

VideoAbuseReportReasonListResponse
There is no detailed description.

VideoAbuseReportReasonMethods
A builder providing access to all methods supported on videoAbuseReportReason resources. It is not used directly, but through the YouTube hub.

VideoAbuseReportReasonSnippet
Basic details about a video category, such as its localized title.

VideoAbuseReportSecondaryReason
There is no detailed description.

VideoAgeGating
There is no detailed description.

VideoCategory
A videoCategory resource identifies a category that has been or could be associated with uploaded videos.

VideoCategoryListCall
Retrieves a list of resources, possibly filtered.

VideoCategoryListResponse
There is no detailed description.

VideoCategoryMethods
A builder providing access to all methods supported on videoCategory resources. It is not used directly, but through the YouTube hub.

VideoCategorySnippet
Basic details about a video category, such as its localized title.

VideoContentDetails
Details about the content of a YouTube Video.

VideoContentDetailsRegionRestriction
DEPRECATED Region restriction of the video.

VideoDeleteCall
Deletes a resource.

VideoFileDetails
Describes original video file properties, including technical details about audio and video streams, but also metadata information like content length, digitization time, or geotagging information.

VideoFileDetailsAudioStream
Information about an audio stream.

VideoFileDetailsVideoStream
Information about a video stream.

VideoGetRatingCall
Retrieves the ratings that the authorized user gave to a list of specified videos.

VideoInsertCall
Inserts a new resource into this collection.

VideoListCall
Retrieves a list of resources, possibly filtered.

VideoListResponse
There is no detailed description.

VideoLiveStreamingDetails
Details about the live streaming metadata.

VideoLocalization
Localized versions of certain video properties (e.g. title).

VideoMethods
A builder providing access to all methods supported on video resources. It is not used directly, but through the YouTube hub.

VideoMonetizationDetails
Details about monetization of a YouTube Video.

VideoPlayer
Player to be used for a video playback.

VideoProcessingDetails
Describes processing status and progress and availability of some other Video resource parts.

VideoProcessingDetailsProcessingProgress
Video processing progress and completion time estimate.

VideoProjectDetails
DEPRECATED. b/157517979: This part was never populated after it was added. However, it sees non-zero traffic because there is generated client code in the wild that refers to it [1]. We keep this field and do NOT remove it because otherwise V3 would return an error when this part gets requested [2]. [1] https://developers.google.com/resources/api-libraries/documentation/youtube/v3/csharp/latest/classGoogle_1_1Apis_1_1YouTube_1_1v3_1_1Data_1_1VideoProjectDetails.html [2] http://google3/video/youtube/src/python/servers/data_api/common.py?l=1565-1569&rcl=344141677

VideoRateCall
Adds a like or dislike rating to a video or removes a rating from a video.

VideoRating
Basic details about rating of a video.

VideoRatingListResponse
There is no detailed description.

VideoRecordingDetails
Recording information associated with the video.

VideoReportAbuseCall
Report abuse for a video.

VideoSnippet
Basic details about a video, including title, description, uploader, thumbnails and category.

VideoStatistics
Statistics about the video, such as the number of times the video was viewed or liked.

VideoStatus
Basic details about a video category, such as its localized title. Next Id: 17

VideoSuggestions
Specifies suggestions on how to improve video content, including encoding hints, tag suggestions, and editor suggestions.

VideoSuggestionsTagSuggestion
A single tag suggestion with it's relevance information.

VideoTopicDetails
Freebase topic information related to the video.

VideoUpdateCall
Updates an existing resource.

WatchSettings
Branding properties for the watch. All deprecated.

WatermarkMethods
A builder providing access to all methods supported on watermark resources. It is not used directly, but through the YouTube hub.

WatermarkSetCall
Allows upload of watermark image and setting it for a channel.

WatermarkUnsetCall
Allows removal of channel watermark.

YouTube
Central instance to access all YouTube related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.