Struct google_youtube3::api::ContentRating

pub struct ContentRating {

    pub acb_rating: Option<String>,
    pub agcom_rating: Option<String>,
    pub anatel_rating: Option<String>,
    pub bbfc_rating: Option<String>,
    pub bfvc_rating: Option<String>,
    pub bmukk_rating: Option<String>,
    pub catv_rating: Option<String>,
    pub catvfr_rating: Option<String>,
    pub cbfc_rating: Option<String>,
    pub ccc_rating: Option<String>,
    pub cce_rating: Option<String>,
    pub chfilm_rating: Option<String>,
    pub chvrs_rating: Option<String>,
    pub cicf_rating: Option<String>,
    pub cna_rating: Option<String>,
    pub cnc_rating: Option<String>,
    pub csa_rating: Option<String>,
    pub cscf_rating: Option<String>,
    pub czfilm_rating: Option<String>,
    pub djctq_rating: Option<String>,
    pub djctq_rating_reasons: Option<Vec<String>>,
    pub ecbmct_rating: Option<String>,
    pub eefilm_rating: Option<String>,
    pub egfilm_rating: Option<String>,
    pub eirin_rating: Option<String>,
    pub fcbm_rating: Option<String>,
    pub fco_rating: Option<String>,
    pub fmoc_rating: Option<String>,
    pub fpb_rating: Option<String>,
    pub fpb_rating_reasons: Option<Vec<String>>,
    pub fsk_rating: Option<String>,
    pub grfilm_rating: Option<String>,
    pub icaa_rating: Option<String>,
    pub ifco_rating: Option<String>,
    pub ilfilm_rating: Option<String>,
    pub incaa_rating: Option<String>,
    pub kfcb_rating: Option<String>,
    pub kijkwijzer_rating: Option<String>,
    pub kmrb_rating: Option<String>,
    pub lsf_rating: Option<String>,
    pub mccaa_rating: Option<String>,
    pub mccyp_rating: Option<String>,
    pub mcst_rating: Option<String>,
    pub mda_rating: Option<String>,
    pub medietilsynet_rating: Option<String>,
    pub meku_rating: Option<String>,
    pub mena_mpaa_rating: Option<String>,
    pub mibac_rating: Option<String>,
    pub moc_rating: Option<String>,
    pub moctw_rating: Option<String>,
    pub mpaa_rating: Option<String>,
    pub mpaat_rating: Option<String>,
    pub mtrcb_rating: Option<String>,
    pub nbc_rating: Option<String>,
    pub nbcpl_rating: Option<String>,
    pub nfrc_rating: Option<String>,
    pub nfvcb_rating: Option<String>,
    pub nkclv_rating: Option<String>,
    pub nmc_rating: Option<String>,
    pub oflc_rating: Option<String>,
    pub pefilm_rating: Option<String>,
    pub rcnof_rating: Option<String>,
    pub resorteviolencia_rating: Option<String>,
    pub rtc_rating: Option<String>,
    pub rte_rating: Option<String>,
    pub russia_rating: Option<String>,
    pub skfilm_rating: Option<String>,
    pub smais_rating: Option<String>,
    pub smsa_rating: Option<String>,
    pub tvpg_rating: Option<String>,
    pub yt_rating: Option<String>,

}

Ratings schemes. The country-specific ratings are mostly for movies and shows. LINT.IfChange

This type is not used in any activity, and only used as part of another schema.

Fields
acb_rating: Option<String>

The video’s Australian Classification Board (ACB) or Australian Communications and Media Authority (ACMA) rating. ACMA ratings are used to classify children’s television programming.

agcom_rating: Option<String>

The video’s rating from Italy’s Autorità per le Garanzie nelle Comunicazioni (AGCOM).

anatel_rating: Option<String>

The video’s Anatel (Asociación Nacional de Televisión) rating for Chilean television.

bbfc_rating: Option<String>

The video’s British Board of Film Classification (BBFC) rating.

bfvc_rating: Option<String>

The video’s rating from Thailand’s Board of Film and Video Censors.

bmukk_rating: Option<String>

The video’s rating from the Austrian Board of Media Classification (Bundesministerium für Unterricht, Kunst und Kultur).

catv_rating: Option<String>

Rating system for Canadian TV - Canadian TV Classification System The video’s rating from the Canadian Radio-Television and Telecommunications Commission (CRTC) for Canadian English-language broadcasts. For more information, see the Canadian Broadcast Standards Council website.

catvfr_rating: Option<String>

The video’s rating from the Canadian Radio-Television and Telecommunications Commission (CRTC) for Canadian French-language broadcasts. For more information, see the Canadian Broadcast Standards Council website.

cbfc_rating: Option<String>

The video’s Central Board of Film Certification (CBFC - India) rating.

ccc_rating: Option<String>

The video’s Consejo de Calificación Cinematográfica (Chile) rating.

cce_rating: Option<String>

The video’s rating from Portugal’s Comissão de Classificação de Espect´culos.

chfilm_rating: Option<String>

The video’s rating in Switzerland.

chvrs_rating: Option<String>

The video’s Canadian Home Video Rating System (CHVRS) rating.

cicf_rating: Option<String>

The video’s rating from the Commission de Contrôle des Films (Belgium).

cna_rating: Option<String>

The video’s rating from Romania’s CONSILIUL NATIONAL AL AUDIOVIZUALULUI (CNA).

cnc_rating: Option<String>

Rating system in France - Commission de classification cinematographique

csa_rating: Option<String>

The video’s rating from France’s Conseil supérieur de l’audiovisuel, which rates broadcast content.

cscf_rating: Option<String>

The video’s rating from Luxembourg’s Commission de surveillance de la classification des films (CSCF).

czfilm_rating: Option<String>

The video’s rating in the Czech Republic.

djctq_rating: Option<String>

The video’s Departamento de Justiça, Classificação, Qualificação e Títulos (DJCQT - Brazil) rating.

djctq_rating_reasons: Option<Vec<String>>

Reasons that explain why the video received its DJCQT (Brazil) rating.

ecbmct_rating: Option<String>

Rating system in Turkey - Evaluation and Classification Board of the Ministry of Culture and Tourism

eefilm_rating: Option<String>

The video’s rating in Estonia.

egfilm_rating: Option<String>

The video’s rating in Egypt.

eirin_rating: Option<String>

The video’s Eirin (映倫) rating. Eirin is the Japanese rating system.

fcbm_rating: Option<String>

The video’s rating from Malaysia’s Film Censorship Board.

fco_rating: Option<String>

The video’s rating from Hong Kong’s Office for Film, Newspaper and Article Administration.

fmoc_rating: Option<String>

This property has been deprecated. Use the contentDetails.contentRating.cncRating instead.

fpb_rating: Option<String>

The video’s rating from South Africa’s Film and Publication Board.

fpb_rating_reasons: Option<Vec<String>>

Reasons that explain why the video received its FPB (South Africa) rating.

fsk_rating: Option<String>

The video’s Freiwillige Selbstkontrolle der Filmwirtschaft (FSK - Germany) rating.

grfilm_rating: Option<String>

The video’s rating in Greece.

icaa_rating: Option<String>

The video’s Instituto de la Cinematografía y de las Artes Audiovisuales (ICAA - Spain) rating.

ifco_rating: Option<String>

The video’s Irish Film Classification Office (IFCO - Ireland) rating. See the IFCO website for more information.

ilfilm_rating: Option<String>

The video’s rating in Israel.

incaa_rating: Option<String>

The video’s INCAA (Instituto Nacional de Cine y Artes Audiovisuales - Argentina) rating.

kfcb_rating: Option<String>

The video’s rating from the Kenya Film Classification Board.

kijkwijzer_rating: Option<String>

The video’s NICAM/Kijkwijzer rating from the Nederlands Instituut voor de Classificatie van Audiovisuele Media (Netherlands).

kmrb_rating: Option<String>

The video’s Korea Media Rating Board (영상물등급위원회) rating. The KMRB rates videos in South Korea.

lsf_rating: Option<String>

The video’s rating from Indonesia’s Lembaga Sensor Film.

mccaa_rating: Option<String>

The video’s rating from Malta’s Film Age-Classification Board.

mccyp_rating: Option<String>

The video’s rating from the Danish Film Institute’s (Det Danske Filminstitut) Media Council for Children and Young People.

mcst_rating: Option<String>

The video’s rating system for Vietnam - MCST

mda_rating: Option<String>

The video’s rating from Singapore’s Media Development Authority (MDA) and, specifically, it’s Board of Film Censors (BFC).

medietilsynet_rating: Option<String>

The video’s rating from Medietilsynet, the Norwegian Media Authority.

meku_rating: Option<String>

The video’s rating from Finland’s Kansallinen Audiovisuaalinen Instituutti (National Audiovisual Institute).

mena_mpaa_rating: Option<String>

The rating system for MENA countries, a clone of MPAA. It is needed to prevent titles go live w/o additional QC check, since some of them can be inappropriate for the countries at all. See b/33408548 for more details.

mibac_rating: Option<String>

The video’s rating from the Ministero dei Beni e delle Attività Culturali e del Turismo (Italy).

moc_rating: Option<String>

The video’s Ministerio de Cultura (Colombia) rating.

moctw_rating: Option<String>

The video’s rating from Taiwan’s Ministry of Culture (文化部).

mpaa_rating: Option<String>

The video’s Motion Picture Association of America (MPAA) rating.

mpaat_rating: Option<String>

The rating system for trailer, DVD, and Ad in the US. See http://movielabs.com/md/ratings/v2.3/html/US_MPAAT_Ratings.html.

mtrcb_rating: Option<String>

The video’s rating from the Movie and Television Review and Classification Board (Philippines).

nbc_rating: Option<String>

The video’s rating from the Maldives National Bureau of Classification.

nbcpl_rating: Option<String>

The video’s rating in Poland.

nfrc_rating: Option<String>

The video’s rating from the Bulgarian National Film Center.

nfvcb_rating: Option<String>

The video’s rating from Nigeria’s National Film and Video Censors Board.

nkclv_rating: Option<String>

The video’s rating from the Nacionãlais Kino centrs (National Film Centre of Latvia).

nmc_rating: Option<String>

The National Media Council ratings system for United Arab Emirates.

oflc_rating: Option<String>

The video’s Office of Film and Literature Classification (OFLC - New Zealand) rating.

pefilm_rating: Option<String>

The video’s rating in Peru.

rcnof_rating: Option<String>

The video’s rating from the Hungarian Nemzeti Filmiroda, the Rating Committee of the National Office of Film.

resorteviolencia_rating: Option<String>

The video’s rating in Venezuela.

rtc_rating: Option<String>

The video’s General Directorate of Radio, Television and Cinematography (Mexico) rating.

rte_rating: Option<String>

The video’s rating from Ireland’s Raidió Teilifís Éireann.

russia_rating: Option<String>

The video’s National Film Registry of the Russian Federation (MKRF - Russia) rating.

skfilm_rating: Option<String>

The video’s rating in Slovakia.

smais_rating: Option<String>

The video’s rating in Iceland.

smsa_rating: Option<String>

The video’s rating from Statens medieråd (Sweden’s National Media Council).

tvpg_rating: Option<String>

The video’s TV Parental Guidelines (TVPG) rating.

yt_rating: Option<String>

A rating that YouTube uses to identify age-restricted content.