Struct google_youtube3::ActivityContentDetails

pub struct ActivityContentDetails {
    pub comment: Option<ActivityContentDetailsComment>,
    pub playlist_item: Option<ActivityContentDetailsPlaylistItem>,
    pub like: Option<ActivityContentDetailsLike>,
    pub promoted_item: Option<ActivityContentDetailsPromotedItem>,
    pub recommendation: Option<ActivityContentDetailsRecommendation>,
    pub favorite: Option<ActivityContentDetailsFavorite>,
    pub upload: Option<ActivityContentDetailsUpload>,
    pub social: Option<ActivityContentDetailsSocial>,
    pub channel_item: Option<ActivityContentDetailsChannelItem>,
    pub bulletin: Option<ActivityContentDetailsBulletin>,
    pub subscription: Option<ActivityContentDetailsSubscription>,
}

Details about the content of an activity: the video that was shared, the channel that was subscribed to, etc.

This type is not used in any activity, and only used as part of another schema.

Fields
comment: Option<ActivityContentDetailsComment>

The comment object contains information about a resource that received a comment. This property is only present if the snippet.type is comment.

playlist_item: Option<ActivityContentDetailsPlaylistItem>

The playlistItem object contains information about a new playlist item. This property is only present if the snippet.type is playlistItem.

like: Option<ActivityContentDetailsLike>

The like object contains information about a resource that received a positive (like) rating. This property is only present if the snippet.type is like.

promoted_item: Option<ActivityContentDetailsPromotedItem>

The promotedItem object contains details about a resource which is being promoted. This property is only present if the snippet.type is promotedItem.

recommendation: Option<ActivityContentDetailsRecommendation>

The recommendation object contains information about a recommended resource. This property is only present if the snippet.type is recommendation.

favorite: Option<ActivityContentDetailsFavorite>

The favorite object contains information about a video that was marked as a favorite video. This property is only present if the snippet.type is favorite.

upload: Option<ActivityContentDetailsUpload>

The upload object contains information about the uploaded video. This property is only present if the snippet.type is upload.

social: Option<ActivityContentDetailsSocial>

The social object contains details about a social network post. This property is only present if the snippet.type is social.

channel_item: Option<ActivityContentDetailsChannelItem>

The channelItem object contains details about a resource which was added to a channel. This property is only present if the snippet.type is channelItem.

bulletin: Option<ActivityContentDetailsBulletin>

The bulletin object contains details about a channel bulletin post. This object is only present if the snippet.type is bulletin.

subscription: Option<ActivityContentDetailsSubscription>

The subscription object contains information about a channel that a user subscribed to. This property is only present if the snippet.type is subscription.