Struct google_youtube3::Channel

pub struct Channel {
    pub status: Option<ChannelStatus>,
    pub invideo_promotion: Option<InvideoPromotion>,
    pub kind: Option<String>,
    pub statistics: Option<ChannelStatistics>,
    pub content_owner_details: Option<ChannelContentOwnerDetails>,
    pub topic_details: Option<ChannelTopicDetails>,
    pub content_details: Option<ChannelContentDetails>,
    pub branding_settings: Option<ChannelBrandingSettings>,
    pub conversion_pings: Option<ChannelConversionPings>,
    pub snippet: Option<ChannelSnippet>,
    pub audit_details: Option<ChannelAuditDetails>,
    pub etag: Option<String>,
    pub id: Option<String>,
    pub localizations: Option<HashMap<String, ChannelLocalization>>,
}

A channel resource contains information about a YouTube channel.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list channels (none)

update channels (request|response)

Fields

status

The status object encapsulates information about the privacy status of the channel.

invideo_promotion

The invideoPromotion object encapsulates information about promotion campaign associated with the channel.

kind

Identifies what kind of resource this is. Value: the fixed string "youtube#channel".

statistics

The statistics object encapsulates statistics for the channel.

content_owner_details

The contentOwnerDetails object encapsulates channel data that is relevant for YouTube Partners linked with the channel.

topic_details

The topicDetails object encapsulates information about Freebase topics associated with the channel.

content_details

The contentDetails object encapsulates information about the channel's content.

branding_settings

The brandingSettings object encapsulates information about the branding of the channel.

conversion_pings

The conversionPings object encapsulates information about conversion pings that need to be respected by the channel.

snippet

The snippet object contains basic details about the channel, such as its title, description, and thumbnail images.

audit_details

The auditionDetails object encapsulates channel data that is relevant for YouTube Partners during the audition process.

etag

Etag of this resource.

id

The ID that YouTube uses to uniquely identify the channel.

localizations

Localizations for different languages