Struct google_youtube3::CommentSnippet

pub struct CommentSnippet {
    pub author_channel_url: Option<String>,
    pub channel_id: Option<String>,
    pub video_id: Option<String>,
    pub published_at: Option<String>,
    pub like_count: Option<u32>,
    pub can_rate: Option<bool>,
    pub updated_at: Option<String>,
    pub author_profile_image_url: Option<String>,
    pub author_googleplus_profile_url: Option<String>,
    pub text_display: Option<String>,
    pub author_display_name: Option<String>,
    pub viewer_rating: Option<String>,
    pub moderation_status: Option<String>,
    pub author_channel_id: Option<String>,
    pub parent_id: Option<String>,
    pub text_original: Option<String>,
}

Basic details about a comment, such as its author and text.

This type is not used in any activity, and only used as part of another schema.

Fields

author_channel_url

Link to the author's YouTube channel, if any.

channel_id

The id of the corresponding YouTube channel. In case of a channel comment this is the channel the comment refers to. In case of a video comment it's the video's channel.

video_id

The ID of the video the comment refers to, if any.

published_at

The date and time when the comment was orignally published. The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

like_count

The total number of likes this comment has received.

can_rate

Whether the current viewer can rate this comment.

updated_at

The date and time when was last updated . The value is specified in ISO 8601 (YYYY-MM-DDThh:mm:ss.sZ) format.

author_profile_image_url

The URL for the avatar of the user who posted the comment.

author_googleplus_profile_url

Link to the author's Google+ profile, if any.

text_display

The comment's text. The format is either plain text or HTML dependent on what has been requested. Even the plain text representation may differ from the text originally posted in that it may replace video links with video titles etc.

author_display_name

The name of the user who posted the comment.

viewer_rating

The rating the viewer has given to this comment. For the time being this will never return RATE_TYPE_DISLIKE and instead return RATE_TYPE_NONE. This may change in the future.

moderation_status

The comment's moderation status. Will not be set if the comments were requested through the id filter.

author_channel_id

The id of the author's YouTube channel, if any.

parent_id

The unique id of the parent comment, only set for replies.

text_original

The comment's original raw text as initially posted or last updated. The original text will only be returned if it is accessible to the viewer, which is only guaranteed if the viewer is the comment's author.