Struct google_youtube3::LiveBroadcastContentDetails

pub struct LiveBroadcastContentDetails {
    pub bound_stream_last_update_time_ms: Option<String>,
    pub projection: Option<String>,
    pub bound_stream_id: Option<String>,
    pub enable_embed: Option<bool>,
    pub enable_closed_captions: Option<bool>,
    pub enable_low_latency: Option<bool>,
    pub start_with_slate: Option<bool>,
    pub closed_captions_type: Option<String>,
    pub enable_content_encryption: Option<bool>,
    pub record_from_start: Option<bool>,
    pub enable_dvr: Option<bool>,
    pub monitor_stream: Option<MonitorStreamInfo>,
}

Detailed settings of a broadcast.

This type is not used in any activity, and only used as part of another schema.

Fields
bound_stream_last_update_time_ms: Option<String>

The date and time that the live stream referenced by boundStreamId was last updated.

projection: Option<String>

The projection format of this broadcast. This defaults to rectangular.

bound_stream_id: Option<String>

This value uniquely identifies the live stream bound to the broadcast.

enable_embed: Option<bool>

This setting indicates whether the broadcast video can be played in an embedded player. If you choose to archive the video (using the enableArchive property), this setting will also apply to the archived video.

enable_closed_captions: Option<bool>

This setting indicates whether HTTP POST closed captioning is enabled for this broadcast. The ingestion URL of the closed captions is returned through the liveStreams API. This is mutually exclusive with using the closed_captions_type property, and is equivalent to setting closed_captions_type to CLOSED_CAPTIONS_HTTP_POST.

enable_low_latency: Option<bool>

Indicates whether this broadcast has low latency enabled.

start_with_slate: Option<bool>

This setting indicates whether the broadcast should automatically begin with an in-stream slate when you update the broadcast's status to live. After updating the status, you then need to send a liveCuepoints.insert request that sets the cuepoint's eventState to end to remove the in-stream slate and make your broadcast stream visible to viewers.

closed_captions_type: Option<String>

no description provided

enable_content_encryption: Option<bool>

This setting indicates whether YouTube should enable content encryption for the broadcast.

record_from_start: Option<bool>

Automatically start recording after the event goes live. The default value for this property is true.

Important: You must also set the enableDvr property's value to true if you want the playback to be available immediately after the broadcast ends. If you set this property's value to true but do not also set the enableDvr property to true, there may be a delay of around one day before the archived video will be available for playback.

enable_dvr: Option<bool>

This setting determines whether viewers can access DVR controls while watching the video. DVR controls enable the viewer to control the video playback experience by pausing, rewinding, or fast forwarding content. The default value for this property is true.

Important: You must set the value to true and also set the enableArchive property's value to true if you want to make playback available immediately after the broadcast ends.

monitor_stream: Option<MonitorStreamInfo>

The monitorStream object contains information about the monitor stream, which the broadcaster can use to review the event content before the broadcast stream is shown publicly.