Struct google_youtube3::Video

pub struct Video {
    pub status: Option<VideoStatus>,
    pub topic_details: Option<VideoTopicDetails>,
    pub monetization_details: Option<VideoMonetizationDetails>,
    pub suggestions: Option<VideoSuggestions>,
    pub age_gating: Option<VideoAgeGating>,
    pub file_details: Option<VideoFileDetails>,
    pub player: Option<VideoPlayer>,
    pub id: Option<String>,
    pub localizations: Option<HashMap<String, VideoLocalization>>,
    pub live_streaming_details: Option<VideoLiveStreamingDetails>,
    pub snippet: Option<VideoSnippet>,
    pub kind: Option<String>,
    pub statistics: Option<VideoStatistics>,
    pub project_details: Option<VideoProjectDetails>,
    pub processing_details: Option<VideoProcessingDetails>,
    pub etag: Option<String>,
    pub content_details: Option<VideoContentDetails>,
    pub recording_details: Option<VideoRecordingDetails>,
}

A video resource represents a YouTube video.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

rate videos (none)

report abuse videos (none)

get rating videos (none)

list videos (none)

insert videos (request|response)

update videos (request|response)

delete videos (none)

Fields
status: Option<VideoStatus>

The status object contains information about the video's uploading, processing, and privacy statuses.

topic_details: Option<VideoTopicDetails>

The topicDetails object encapsulates information about Freebase topics associated with the video.

monetization_details: Option<VideoMonetizationDetails>

The monetizationDetails object encapsulates information about the monetization status of the video.

suggestions: Option<VideoSuggestions>

The suggestions object encapsulates suggestions that identify opportunities to improve the video quality or the metadata for the uploaded video. This data can only be retrieved by the video owner.

age_gating: Option<VideoAgeGating>

Age restriction details related to a video. This data can only be retrieved by the video owner.

file_details: Option<VideoFileDetails>

The fileDetails object encapsulates information about the video file that was uploaded to YouTube, including the file's resolution, duration, audio and video codecs, stream bitrates, and more. This data can only be retrieved by the video owner.

player: Option<VideoPlayer>

The player object contains information that you would use to play the video in an embedded player.

id: Option<String>

The ID that YouTube uses to uniquely identify the video.

localizations: Option<HashMap<String, VideoLocalization>>

List with all localizations.

live_streaming_details: Option<VideoLiveStreamingDetails>

The liveStreamingDetails object contains metadata about a live video broadcast. The object will only be present in a video resource if the video is an upcoming, live, or completed live broadcast.

snippet: Option<VideoSnippet>

The snippet object contains basic details about the video, such as its title, description, and category.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "youtube#video".

statistics: Option<VideoStatistics>

The statistics object contains statistics about the video.

project_details: Option<VideoProjectDetails>

The projectDetails object contains information about the project specific video metadata.

processing_details: Option<VideoProcessingDetails>

The processingProgress object encapsulates information about YouTube's progress in processing the uploaded video file. The properties in the object identify the current processing status and an estimate of the time remaining until YouTube finishes processing the video. This part also indicates whether different types of data or content, such as file details or thumbnail images, are available for the video.

The processingProgress object is designed to be polled so that the video uploaded can track the progress that YouTube has made in processing the uploaded video file. This data can only be retrieved by the video owner.

etag: Option<String>

Etag of this resource.

content_details: Option<VideoContentDetails>

The contentDetails object contains information about the video content, including the length of the video and its aspect ratio.

recording_details: Option<VideoRecordingDetails>

The recordingDetails object encapsulates information about the location, date and address where the video was recorded.