Struct google_youtube3::VideoContentDetails

pub struct VideoContentDetails {
    pub definition: Option<String>,
    pub country_restriction: Option<AccessPolicy>,
    pub content_rating: Option<ContentRating>,
    pub caption: Option<String>,
    pub region_restriction: Option<VideoContentDetailsRegionRestriction>,
    pub duration: Option<String>,
    pub licensed_content: Option<bool>,
    pub dimension: Option<String>,
}

Details about the content of a YouTube Video.

This type is not used in any activity, and only used as part of another schema.

Fields

definition

The value of definition indicates whether the video is available in high definition or only in standard definition.

country_restriction

The countryRestriction object contains information about the countries where a video is (or is not) viewable.

content_rating

Specifies the ratings that the video received under various rating schemes.

caption

The value of captions indicates whether the video has captions or not.

region_restriction

The regionRestriction object contains information about the countries where a video is (or is not) viewable. The object will contain either the contentDetails.regionRestriction.allowed property or the contentDetails.regionRestriction.blocked property.

duration

The length of the video. The tag value is an ISO 8601 duration in the format PT#M#S, in which the letters PT indicate that the value specifies a period of time, and the letters M and S refer to length in minutes and seconds, respectively. The # characters preceding the M and S letters are both integers that specify the number of minutes (or seconds) of the video. For example, a value of PT15M51S indicates that the video is 15 minutes and 51 seconds long.

licensed_content

The value of is_license_content indicates whether the video is licensed content.

dimension

The value of dimension indicates whether the video is available in 3D or in 2D.