Struct google_youtube3::YouTube

pub struct YouTube<C, A> {
    // some fields omitted
}

Central instance to access all YouTube related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtube3 as youtube3;
use youtube3::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtube3::YouTube;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTube::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.videos().list("part")
             .video_category_id("labore")
             .region_code("sea")
             .page_token("nonumy")
             .on_behalf_of_content_owner("dolores")
             .my_rating("gubergren")
             .max_results(6)
             .locale("aliquyam")
             .id("ea")
             .hl("no")
             .chart("justo")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> YouTube<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> YouTube<C, A>

fn activities(&'a self) -> ActivityMethods<'a, C, A>

fn captions(&'a self) -> CaptionMethods<'a, C, A>

fn channel_banners(&'a self) -> ChannelBannerMethods<'a, C, A>

fn channel_sections(&'a self) -> ChannelSectionMethods<'a, C, A>

fn channels(&'a self) -> ChannelMethods<'a, C, A>

fn comment_threads(&'a self) -> CommentThreadMethods<'a, C, A>

fn comments(&'a self) -> CommentMethods<'a, C, A>

fn fan_funding_events(&'a self) -> FanFundingEventMethods<'a, C, A>

fn guide_categories(&'a self) -> GuideCategoryMethods<'a, C, A>

fn i18n_languages(&'a self) -> I18nLanguageMethods<'a, C, A>

fn i18n_regions(&'a self) -> I18nRegionMethods<'a, C, A>

fn live_broadcasts(&'a self) -> LiveBroadcastMethods<'a, C, A>

fn live_chat_bans(&'a self) -> LiveChatBanMethods<'a, C, A>

fn live_chat_messages(&'a self) -> LiveChatMessageMethods<'a, C, A>

fn live_chat_moderators(&'a self) -> LiveChatModeratorMethods<'a, C, A>

fn live_streams(&'a self) -> LiveStreamMethods<'a, C, A>

fn playlist_items(&'a self) -> PlaylistItemMethods<'a, C, A>

fn playlists(&'a self) -> PlaylistMethods<'a, C, A>

fn search(&'a self) -> SearchMethods<'a, C, A>

fn sponsors(&'a self) -> SponsorMethods<'a, C, A>

fn subscriptions(&'a self) -> SubscriptionMethods<'a, C, A>

fn thumbnails(&'a self) -> ThumbnailMethods<'a, C, A>

fn video_abuse_report_reasons(&'a self) -> VideoAbuseReportReasonMethods<'a, C, A>

fn video_categories(&'a self) -> VideoCategoryMethods<'a, C, A>

fn videos(&'a self) -> VideoMethods<'a, C, A>

fn watermarks(&'a self) -> WatermarkMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.14.

Returns the previously set user-agent.